Split XServerDrawerMenu.kt pane content into per-pane files (behavior-neutral, 6.2k->2.7k lines)
