+++ app/src/main/runtime/display/XServerDrawerMenu.kt
@@ -190,13 +190,13 @@ private const val DrawerSurfaceAlpha = 0.72f
 private const val DrawerPressedAlpha = 0.88f
 private const val DrawerGradientLift = 0.014f
 
-private val DrawerAccent = Color(0xFF2196F3)
+internal val DrawerAccent = Color(0xFF2196F3)
 private val DrawerActiveAccent = Color(0xFF29B6F6)
 private val DrawerFocusFill = Color(0xFF0E2438)
-private val DrawerTextPrimary = WinNativeTextPrimary.copy(alpha = 0.88f)
-private val DrawerTextSecondary = WinNativeTextSecondary.copy(alpha = 0.82f)
-private val DrawerOutline = WinNativeOutline
-private val DrawerBackground = WinNativeBackground.copy(alpha = DrawerSheetAlpha)
+internal val DrawerTextPrimary = WinNativeTextPrimary.copy(alpha = 0.88f)
+internal val DrawerTextSecondary = WinNativeTextSecondary.copy(alpha = 0.82f)
+internal val DrawerOutline = WinNativeOutline
+internal val DrawerBackground = WinNativeBackground.copy(alpha = DrawerSheetAlpha)
 
 internal val PaneSurfaceColor = WinNativeBackground.copy(alpha = DrawerSheetAlpha)
 private val PaneSurfacePressed = Color(0xFF232B3A).copy(alpha = DrawerPressedAlpha)
@@ -204,19 +204,19 @@ private val PaneSurfacePressed = Color(0xFF232B3A).copy(alpha = DrawerPressedAlp
 private val TopRailSurfaceColor = WinNativeSurface.copy(alpha = DrawerSheetAlpha)
 
 private val TileResting = Color(0xFF20283A).copy(alpha = DrawerSurfaceAlpha)
-private val TileExitResting = Color(0xFF3A2125).copy(alpha = DrawerSurfaceAlpha)
-private val TileExitPressed = Color(0xFF4A2A30).copy(alpha = DrawerPressedAlpha)
-private val PaneInnerResting = WinNativePanel.copy(alpha = DrawerSurfaceAlpha)
-private val PaneInnerPressed = Color(0xFF242B3A).copy(alpha = DrawerPressedAlpha)
-private val RestingCardBorder = WinNativeOutline.copy(alpha = 0.72f)
+internal val TileExitResting = Color(0xFF3A2125).copy(alpha = DrawerSurfaceAlpha)
+internal val TileExitPressed = Color(0xFF4A2A30).copy(alpha = DrawerPressedAlpha)
+internal val PaneInnerResting = WinNativePanel.copy(alpha = DrawerSurfaceAlpha)
+internal val PaneInnerPressed = Color(0xFF242B3A).copy(alpha = DrawerPressedAlpha)
+internal val RestingCardBorder = WinNativeOutline.copy(alpha = 0.72f)
 private val DisabledCardBorder = Color(0xFF202033).copy(alpha = 0.58f)
-private val ActiveCardBorder = DrawerActiveAccent
+internal val ActiveCardBorder = DrawerActiveAccent
 private val BottomDividerColor = WinNativeOutline
-private val GlassExitTint = Color(0xFFE07B6B)
-private val RecordRed = Color(0xFFE53935)
+internal val GlassExitTint = Color(0xFFE07B6B)
+internal val RecordRed = Color(0xFFE53935)
 
 // Pane content scales down on short displays.
-private val LocalPaneScale = staticCompositionLocalOf { 1f }
+internal val LocalPaneScale = staticCompositionLocalOf { 1f }
 
 private const val PANE_DIR_LEFT = 1
 private const val PANE_DIR_RIGHT = 2
@@ -235,7 +235,7 @@ private class PaneNavEntry(
 )
 
 @Stable
-private class PaneNavRegistry {
+internal class PaneNavRegistry {
     private val items = mutableStateMapOf<Int, PaneNavEntry>()
     private var slotCounter = 0
     private var lastSignal = -1
@@ -334,9 +334,9 @@ private class PaneNavRegistry {
     }
 }
 
-private val LocalPaneNav = staticCompositionLocalOf<PaneNavRegistry?> { null }
+internal val LocalPaneNav = staticCompositionLocalOf<PaneNavRegistry?> { null }
 
-private fun Modifier.paneHighlight(highlighted: Boolean, cornerRadius: Dp): Modifier =
+internal fun Modifier.paneHighlight(highlighted: Boolean, cornerRadius: Dp): Modifier =
     drawWithContent {
         val cr = cornerRadius.toPx()
         if (highlighted) {
@@ -354,7 +354,7 @@ private fun Modifier.paneHighlight(highlighted: Boolean, cornerRadius: Dp): Modi
 
 @OptIn(ExperimentalFoundationApi::class)
 @Composable
-private fun Modifier.paneNavItem(
+internal fun Modifier.paneNavItem(
     cornerRadius: Dp = 10.dp,
     onActivate: () -> Unit = {},
     onAdjust: (Int) -> Unit = {},
@@ -379,7 +379,7 @@ private fun Modifier.paneNavItem(
 }
 
 @Composable
-private fun NavBooleanRow(
+internal fun NavBooleanRow(
     title: String,
     checked: Boolean,
     onCheckedChange: (Boolean) -> Unit,
@@ -397,7 +397,7 @@ private fun NavBooleanRow(
 }
 
 @Composable
-private fun NavEnableRow(
+internal fun NavEnableRow(
     title: String,
     checked: Boolean,
     onCheckedChange: (Boolean) -> Unit,
@@ -414,7 +414,7 @@ private fun NavEnableRow(
 }
 
 @Composable
-private fun NavSliderRow(
+internal fun NavSliderRow(
     label: String,
     valueText: String,
     value: Float,
@@ -459,14 +459,14 @@ private fun NavSliderRow(
 }
 
 private const val PaneScaleMin = 0.78f
-private const val ControlsPaneScaleMin = 0.62f
+internal const val ControlsPaneScaleMin = 0.62f
 private const val PaneScaleReferenceHeightDp = 520f
-private const val PendingTaskAffinityTimeoutMs = 2500L
+internal const val PendingTaskAffinityTimeoutMs = 2500L
 
-private fun computePaneScale(availableHeight: Dp, minScale: Float = PaneScaleMin): Float =
+internal fun computePaneScale(availableHeight: Dp, minScale: Float = PaneScaleMin): Float =
     (availableHeight.value / PaneScaleReferenceHeightDp).coerceIn(minScale, 1f)
 
-private enum class HUDMetricEditor(
+internal enum class HUDMetricEditor(
     val minPercent: Int,
     val maxPercent: Int,
 ) {
@@ -502,7 +502,7 @@ data class TaskManagerPaneState(
     val memoryDetail: String = "",
 )
 
-private data class PendingTaskAffinity(
+internal data class PendingTaskAffinity(
     val affinityMask: Int,
     val requestedAtMillis: Long,
 )
@@ -2295,7 +2295,7 @@ private fun BottomActionButton(
 }
 
 @Composable
-private fun ThinDivider() {
+internal fun ThinDivider() {
     Box(
         modifier =
             Modifier
@@ -2306,7 +2306,7 @@ private fun ThinDivider() {
 }
 
 @Composable
-private fun DrawerResetRow(label: String, onClick: () -> Unit) {
+internal fun DrawerResetRow(label: String, onClick: () -> Unit) {
     val paneScale = LocalPaneScale.current
     val shape = RoundedCornerShape((14f * paneScale).dp)
     Row(
@@ -2358,2361 +2358,8 @@ private fun PaneEnableRow(
     )
 }
 
-@Composable
-private fun HUDPaneContent(
-    state: XServerDrawerState,
-    listener: XServerDrawerActionListener,
-) {
-    var activeEditor by remember { mutableStateOf<HUDMetricEditor?>(null) }
-    var fpsLimitMemory by remember {
-        mutableStateOf(if (state.fpsLimit > 0) state.fpsLimit else FPS_LIMITER_DEFAULT)
-    }
-    LaunchedEffect(state.fpsLimit) {
-        if (state.fpsLimit > 0) fpsLimitMemory = state.fpsLimit
-    }
-    val elementNames =
-        listOf(
-            stringResource(R.string.session_drawer_hud_element_fps),
-            stringResource(R.string.session_drawer_hud_element_api),
-            stringResource(R.string.session_drawer_hud_element_gpu),
-            stringResource(R.string.session_drawer_hud_element_cpu),
-            stringResource(R.string.session_drawer_hud_element_ram),
-            stringResource(R.string.session_drawer_hud_element_battery),
-            stringResource(R.string.session_drawer_hud_element_temp),
-            stringResource(R.string.session_drawer_hud_element_graph),
-            stringResource(R.string.session_drawer_hud_element_cpu_temp),
-        )
-    val elementOrder = listOf(1, 2, 3, 8, 4, 5, 6, 0, 7)
-    val active =
-        state.items.firstOrNull { it.itemId == R.id.main_menu_fps_monitor }?.active ?: false
 
-    activeEditor?.let { editor ->
-        HUDMetricInputDialog(
-            editor = editor,
-            initialPercent =
-                when (editor) {
-                    HUDMetricEditor.ALPHA -> (state.hudTransparency * 100).roundToInt()
-                    HUDMetricEditor.BACKGROUND_ALPHA -> (state.hudBackgroundTransparency * 100).roundToInt()
-                    HUDMetricEditor.SCALE -> (state.hudScale * 100).roundToInt()
-                },
-            onDismiss = { activeEditor = null },
-            onConfirm = { enteredPercent ->
-                activeEditor = null
-                when (editor) {
-                    HUDMetricEditor.ALPHA -> {
-                        listener.onHUDTransparencyChanged(enteredPercent.coerceIn(editor.minPercent, editor.maxPercent) / 100f)
-                    }
-                    HUDMetricEditor.BACKGROUND_ALPHA -> {
-                        listener.onHUDBackgroundTransparencyChanged(enteredPercent.coerceIn(editor.minPercent, editor.maxPercent) / 100f)
-                    }
-                    HUDMetricEditor.SCALE -> {
-                        listener.onHUDScaleChanged(enteredPercent.coerceIn(editor.minPercent, editor.maxPercent) / 100f)
-                    }
-                }
-            },
-        )
-    }
-
-    BoxWithConstraints(modifier = Modifier.fillMaxSize()) {
-        val paneScale = computePaneScale(maxHeight)
-        CompositionLocalProvider(LocalPaneScale provides paneScale) {
-            Column(
-                modifier =
-                    Modifier
-                        .fillMaxWidth()
-                        .verticalScroll(rememberScrollState())
-                        .padding(horizontal = (12f * paneScale).dp, vertical = (12f * paneScale).dp),
-                verticalArrangement = Arrangement.spacedBy((10f * paneScale).dp),
-            ) {
-            NavEnableRow(
-                title = stringResource(R.string.session_drawer_fps_monitor),
-                checked = active,
-                onCheckedChange = { listener.onActionSelected(R.id.main_menu_fps_monitor) },
-            )
-
-            if (active) {
-                NavSliderRow(
-                    label = stringResource(R.string.session_drawer_hud_alpha),
-                    valueText = "${(state.hudTransparency * 100).toInt()}%",
-                    value = state.hudTransparency,
-                    valueRange = 0.1f..1f,
-                    steps = 17,
-                    onValueClick = { activeEditor = HUDMetricEditor.ALPHA },
-                    onValueChange = { listener.onHUDTransparencyChanged(it.snapToStep(0.05f, 0.1f, 1f)) },
-                )
-
-                if (state.hudBackgroundAlphaEnabled) {
-                    NavSliderRow(
-                        label = stringResource(R.string.session_drawer_hud_background),
-                        valueText = "${(state.hudBackgroundTransparency * 100).toInt()}%",
-                        value = state.hudBackgroundTransparency,
-                        valueRange = 0.1f..1f,
-                        steps = 17,
-                        onValueClick = { activeEditor = HUDMetricEditor.BACKGROUND_ALPHA },
-                        onValueChange = { listener.onHUDBackgroundTransparencyChanged(it.snapToStep(0.05f, 0.1f, 1f)) },
-                    )
-                }
-
-                NavSliderRow(
-                    label = stringResource(R.string.session_drawer_hud_scale),
-                    valueText = "${Math.round(state.hudScale * 100)}%",
-                    value = state.hudScale,
-                    valueRange = 0.3f..2.0f,
-                    steps = 33,
-                    onValueClick = { activeEditor = HUDMetricEditor.SCALE },
-                    onValueChange = { listener.onHUDScaleChanged(it.snapToStep(0.05f, 0.3f, 2.0f)) },
-                    adjustStep = 0.05f,
-                )
-
-                NavBooleanRow(
-                    title = stringResource(R.string.session_drawer_hud_background_alpha),
-                    checked = state.hudBackgroundAlphaEnabled,
-                    onCheckedChange = listener::onHUDBackgroundAlphaDecoupledChanged,
-                )
-
-                NavBooleanRow(
-                    title = stringResource(R.string.session_drawer_hud_frametime_numeric),
-                    checked = state.frametimeNumericEnabled,
-                    onCheckedChange = listener::onFrametimeNumericChanged,
-                )
-
-                Box(
-                    Modifier.fillMaxWidth().paneNavItem(
-                        cornerRadius = (12f * paneScale).dp,
-                        onActivate = { listener.onFPSLimitChanged(if (state.fpsLimit > 0) 0 else fpsLimitMemory.coerceIn(FPS_LIMITER_MIN, state.maxRefreshRate)) },
-                        onAdjust = { dir ->
-                            val base = if (state.fpsLimit > 0) state.fpsLimit else fpsLimitMemory
-                            val q = base / 5.0
-                            val units = if (dir > 0) Math.floor(q + 1e-4) + 1 else Math.ceil(q - 1e-4) - 1
-                            listener.onFPSLimitChanged((units * 5).toInt().coerceIn(FPS_LIMITER_MIN, state.maxRefreshRate))
-                        },
-                    ),
-                ) {
-                    FPSLimiterCard(
-                        currentLimit = state.fpsLimit,
-                        maxRefreshRate = state.maxRefreshRate,
-                        onLimitChanged = listener::onFPSLimitChanged,
-                    )
-                }
-
-                Column(verticalArrangement = Arrangement.spacedBy((8f * paneScale).dp)) {
-                    PaneSectionLabel(stringResource(R.string.session_drawer_hud_elements))
-                    ChipFlow {
-                        elementOrder.forEach { index ->
-                            HUDToggleChip(
-                                label = elementNames[index],
-                                checked = state.hudElements[index],
-                                onClick = { listener.onHUDElementToggled(index, !state.hudElements[index]) },
-                                modifier = Modifier.paneNavItem(
-                                    cornerRadius = (16f * paneScale).dp,
-                                    onActivate = { listener.onHUDElementToggled(index, !state.hudElements[index]) },
-                                ),
-                            )
-                        }
-                    }
-                }
-
-                NavBooleanRow(
-                    title = stringResource(R.string.session_drawer_dual_series_battery),
-                    checked = state.dualSeriesBatteryEnabled,
-                    onCheckedChange = listener::onDualSeriesBatteryChanged,
-                )
-            }
-            }
-        }
-    }
-}
-
-@Composable
-private fun TouchPaneContent(
-    state: XServerDrawerState,
-    listener: XServerDrawerActionListener,
-    onClose: () -> Unit,
-) {
-    BoxWithConstraints(modifier = Modifier.fillMaxSize()) {
-        val paneScale = computePaneScale(maxHeight)
-        CompositionLocalProvider(LocalPaneScale provides paneScale) {
-            Column(
-                modifier =
-                    Modifier
-                        .fillMaxWidth()
-                        .verticalScroll(rememberScrollState())
-                        .padding(horizontal = (12f * paneScale).dp, vertical = (12f * paneScale).dp),
-                verticalArrangement = Arrangement.spacedBy((10f * paneScale).dp),
-            ) {
-                Box(
-                    modifier =
-                        Modifier
-                            .size((30f * paneScale).dp)
-                            .clip(RoundedCornerShape((8f * paneScale).dp))
-                            .paneNavItem(cornerRadius = (8f * paneScale).dp, onActivate = onClose)
-                            .clickable(
-                                interactionSource = remember { MutableInteractionSource() },
-                                indication = null,
-                                onClick = onClose,
-                            ),
-                    contentAlignment = Alignment.Center,
-                ) {
-                    Icon(
-                        imageVector = Icons.AutoMirrored.Outlined.ArrowBack,
-                        contentDescription = stringResource(R.string.common_ui_back),
-                        tint = DrawerTextSecondary,
-                        modifier = Modifier.size((20f * paneScale).dp),
-                    )
-                }
-                Box(Modifier.fillMaxWidth().paneNavItem(cornerRadius = (12f * paneScale).dp, onActivate = { listener.onActionSelected(R.id.main_menu_disable_mouse) })) {
-                    DrawerBooleanRow(
-                        title = stringResource(R.string.session_drawer_mouse_input),
-                        checked = state.mouseEnabled,
-                        onCheckedChange = { listener.onActionSelected(R.id.main_menu_disable_mouse) },
-                    )
-                }
-                Box(Modifier.fillMaxWidth().paneNavItem(cornerRadius = (12f * paneScale).dp, onActivate = { listener.onActionSelected(R.id.main_menu_relative_mouse_movement) })) {
-                    DrawerBooleanRow(
-                        title = stringResource(R.string.session_drawer_relative_mouse_movement),
-                        checked = state.relativeMouseEnabled,
-                        onCheckedChange = { listener.onActionSelected(R.id.main_menu_relative_mouse_movement) },
-                    )
-                }
-                Box(Modifier.fillMaxWidth().paneNavItem(cornerRadius = (12f * paneScale).dp, onActivate = { listener.onScreenTouchModeChanged(0) })) {
-                    DrawerBooleanRow(
-                        title = stringResource(R.string.session_drawer_touch_trackpad),
-                        checked = state.screenTouchMode == 0 && !state.rtsGesturesEnabled,
-                        onCheckedChange = { if (it) listener.onScreenTouchModeChanged(0) },
-                    )
-                }
-                Box(Modifier.fillMaxWidth().paneNavItem(cornerRadius = (12f * paneScale).dp, onActivate = { listener.onScreenTouchModeChanged(1) })) {
-                    DrawerBooleanRow(
-                        title = stringResource(R.string.session_drawer_touch_touchscreen),
-                        checked = state.screenTouchMode == 1 && !state.rtsGesturesEnabled,
-                        onCheckedChange = { listener.onScreenTouchModeChanged(if (it) 1 else 0) },
-                    )
-                }
-                Box(Modifier.fillMaxWidth().paneNavItem(cornerRadius = (12f * paneScale).dp, onActivate = { listener.onScreenTouchModeChanged(2) })) {
-                    DrawerBooleanRow(
-                        title = stringResource(R.string.session_drawer_touch_map_right_stick),
-                        checked = state.screenTouchMode == 2 && !state.rtsGesturesEnabled,
-                        onCheckedChange = { listener.onScreenTouchModeChanged(if (it) 2 else 0) },
-                    )
-                }
-                Box(Modifier.fillMaxWidth().paneNavItem(cornerRadius = (12f * paneScale).dp, onActivate = { listener.onRtsGesturesToggled(!state.rtsGesturesEnabled) })) {
-                    DrawerBooleanRow(
-                        title = stringResource(R.string.session_drawer_rts_gestures),
-                        checked = state.rtsGesturesEnabled,
-                        onCheckedChange = { listener.onRtsGesturesToggled(it) },
-                    )
-                }
-                if (state.rtsGesturesEnabled) {
-                    Column(verticalArrangement = Arrangement.spacedBy((8f * paneScale).dp)) {
-                        PaneSectionLabel(stringResource(R.string.session_gesture_profile_section))
-                        InputControlsProfileSelector(
-                            profileNames = state.gestureProfileNames,
-                            selectedIndex = state.gestureSelectedProfileIndex,
-                            onProfileSelected = listener::onGestureProfileSelected,
-                            onEditClick = listener::onRtsGesturesEditClick,
-                        )
-                    }
-                }
-            }
-        }
-    }
-}
-
-@Composable
-private fun GyroscopePaneContent(
-    state: XServerDrawerState,
-    listener: XServerDrawerActionListener,
-) {
-    var calibrateExpanded by remember { mutableStateOf(false) }
-
-    BoxWithConstraints(modifier = Modifier.fillMaxSize()) {
-        val paneScale = computePaneScale(maxHeight)
-        CompositionLocalProvider(LocalPaneScale provides paneScale) {
-            Column(
-                modifier =
-                    Modifier
-                        .fillMaxWidth()
-                        .verticalScroll(rememberScrollState())
-                        .padding(horizontal = (12f * paneScale).dp, vertical = (12f * paneScale).dp),
-                verticalArrangement = Arrangement.spacedBy((10f * paneScale).dp),
-            ) {
-            NavEnableRow(
-                title = stringResource(R.string.session_gyroscope_title),
-                checked = state.gyroscopeEnabled,
-                onCheckedChange = listener::onGyroscopeEnabledChanged,
-            )
-
-            if (state.gyroscopeEnabled) {
-                Column(verticalArrangement = Arrangement.spacedBy((8f * paneScale).dp)) {
-                    PaneSectionLabel(stringResource(R.string.session_gyroscope_mode))
-                    Row(
-                        modifier = Modifier.fillMaxWidth(),
-                        horizontalArrangement = Arrangement.spacedBy((8f * paneScale).dp),
-                    ) {
-                        listOf(
-                            stringResource(R.string.session_gyroscope_hold),
-                            stringResource(R.string.session_gyroscope_toggle),
-                        ).forEachIndexed { index, label ->
-                            HUDToggleChip(
-                                label = label,
-                                checked = state.gyroscopeModeIndex == index,
-                                onClick = { listener.onGyroscopeModeSelected(index) },
-                                modifier = Modifier.weight(1f).paneNavItem(
-                                    cornerRadius = (16f * paneScale).dp,
-                                    onActivate = { listener.onGyroscopeModeSelected(index) },
-                                ),
-                            )
-                        }
-                    }
-                }
-
-                NavBooleanRow(
-                    title = stringResource(R.string.session_gyroscope_orientation_mode),
-                    checked = state.gyroOrientationEnabled,
-                    onCheckedChange = listener::onGyroOrientationModeChanged,
-                )
-
-                Column(verticalArrangement = Arrangement.spacedBy((8f * paneScale).dp)) {
-                    PaneSectionLabel(stringResource(R.string.session_gyroscope_activator_button))
-                    GyroscopeActivatorDropdown(
-                        currentLabel = state.gyroscopeActivatorLabel,
-                        onSelected = listener::onGyroscopeActivatorSelected,
-                    )
-                }
-
-                NavBooleanRow(
-                    title = stringResource(R.string.session_gyroscope_enable_right_stick),
-                    checked = state.rightStickGyroEnabled,
-                    onCheckedChange = listener::onRightStickGyroChanged,
-                )
-
-                NavBooleanRow(
-                    title = stringResource(R.string.session_gyroscope_experimental_mouse_movement),
-                    checked = state.gyroMouseEnabled,
-                    onCheckedChange = listener::onGyroMouseEnabledChanged,
-                )
-
-                if (state.gyroMouseEnabled) {
-                    NavSliderRow(
-                        label = stringResource(R.string.session_gyroscope_mouse_scale),
-                        valueText = "${state.gyroMouseScale.toInt()}%",
-                        value = state.gyroMouseScale,
-                        valueRange = 0f..200f,
-                        steps = 199,
-                        onValueChange = { listener.onGyroMouseScaleChanged(it.roundToInt().toFloat()) },
-                    )
-                }
-
-                ExpandableSection(
-                    title = stringResource(R.string.session_drawer_calibrate_advanced),
-                    expanded = calibrateExpanded,
-                    onToggle = { calibrateExpanded = !calibrateExpanded },
-                ) {
-                    NavSliderRow(
-                        label = stringResource(R.string.session_gyroscope_x_sensitivity),
-                        valueText = "${(state.gyroXSensitivity * 100).roundToInt()}%",
-                        value = state.gyroXSensitivity,
-                        valueRange = 0.01f..3f,
-                        steps = 0,
-                        onValueChange = { listener.onGyroXSensitivityChanged(it) },
-                    )
-
-                    NavSliderRow(
-                        label = stringResource(R.string.session_gyroscope_y_sensitivity),
-                        valueText = "${(state.gyroYSensitivity * 100).roundToInt()}%",
-                        value = state.gyroYSensitivity,
-                        valueRange = 0.01f..3f,
-                        steps = 0,
-                        onValueChange = { listener.onGyroYSensitivityChanged(it) },
-                    )
-
-                    NavSliderRow(
-                        label = stringResource(R.string.session_gyroscope_smoothing),
-                        valueText = "${(state.gyroSmoothing * 100).toInt()}%",
-                        value = state.gyroSmoothing,
-                        valueRange = 0f..1f,
-                        steps = 99,
-                        onValueChange = { listener.onGyroSmoothingChanged(it) },
-                    )
-
-                    NavSliderRow(
-                        label = stringResource(R.string.session_gyroscope_deadzone),
-                        valueText = "${(state.gyroDeadzone * 100).toInt()}%",
-                        value = state.gyroDeadzone,
-                        valueRange = 0f..1f,
-                        steps = 99,
-                        onValueChange = { listener.onGyroDeadzoneChanged(it) },
-                    )
-
-                    Row(
-                        modifier = Modifier.fillMaxWidth(),
-                        horizontalArrangement = Arrangement.spacedBy((8f * paneScale).dp),
-                    ) {
-                        HUDToggleChip(
-                            label = stringResource(R.string.session_gyroscope_invert_x),
-                            checked = state.invertGyroX,
-                            onClick = { listener.onInvertGyroXChanged(!state.invertGyroX) },
-                            modifier = Modifier.weight(1f).paneNavItem(cornerRadius = (16f * paneScale).dp, onActivate = { listener.onInvertGyroXChanged(!state.invertGyroX) }),
-                        )
-                        HUDToggleChip(
-                            label = stringResource(R.string.session_gyroscope_invert_y),
-                            checked = state.invertGyroY,
-                            onClick = { listener.onInvertGyroYChanged(!state.invertGyroY) },
-                            modifier = Modifier.weight(1f).paneNavItem(cornerRadius = (16f * paneScale).dp, onActivate = { listener.onInvertGyroYChanged(!state.invertGyroY) }),
-                        )
-                    }
-
-                    Box(
-                        Modifier.paneNavItem(
-                            cornerRadius = 10.dp,
-                            onActivate = { listener.onActionSelected(R.id.main_menu_gyroscope_reset) },
-                        ),
-                    ) {
-                        WinNativeDialogButton(
-                            label = stringResource(R.string.session_gyroscope_reset_stick),
-                            textColor = DrawerAccent,
-                            backgroundColor = DrawerAccent.copy(alpha = 0.12f),
-                            borderColor = DrawerAccent.copy(alpha = 0.3f),
-                            onClick = { listener.onActionSelected(R.id.main_menu_gyroscope_reset) },
-                        )
-                    }
-                }
-            }
-            }
-        }
-    }
-}
-
-@Composable
-private fun InputControlsPaneContent(
-    state: XServerDrawerState,
-    listener: XServerDrawerActionListener,
-) {
-    BoxWithConstraints(modifier = Modifier.fillMaxSize()) {
-        val paneScale = computePaneScale(maxHeight, ControlsPaneScaleMin)
-        val scrollState = rememberScrollState()
-        val gcmEnabled = state.inputControlsGcmRumbleMode != "disabled"
-        CompositionLocalProvider(LocalPaneScale provides paneScale) {
-            Column(
-                modifier =
-                    Modifier
-                        .fillMaxWidth()
-                        .verticalScroll(scrollState)
-                        .padding(
-                            start = (12f * paneScale).dp,
-                            end = (12f * paneScale).dp,
-                            top = (4f * paneScale).dp,
-                            bottom = (12f * paneScale).dp,
-                        ),
-                verticalArrangement = Arrangement.spacedBy((10f * paneScale).dp),
-            ) {
-                Column(verticalArrangement = Arrangement.spacedBy((8f * paneScale).dp)) {
-                    PaneSectionLabel(stringResource(R.string.input_controls_editor_select_profile))
-                    InputControlsProfileSelector(
-                        profileNames = state.inputControlsProfileNames,
-                        selectedIndex = state.inputControlsSelectedProfileIndex,
-                        onProfileSelected = listener::onInputControlsProfileSelected,
-                        onEditClick = listener::onInputControlsEditClick,
-                    )
-                }
-
-                if (state.inputControlsStyleNames.isNotEmpty()) {
-                    Column(verticalArrangement = Arrangement.spacedBy((8f * paneScale).dp)) {
-                        PaneSectionLabel(stringResource(R.string.input_controls_select_style))
-                        InputControlsSimpleDropdown(
-                            options = state.inputControlsStyleNames,
-                            selectedIndex = state.inputControlsSelectedStyleIndex,
-                            onSelected = listener::onInputControlsStyleSelected,
-                        )
-                    }
-                }
-
-                if (state.inputControlsLabelThemeNames.isNotEmpty()) {
-                    Column(verticalArrangement = Arrangement.spacedBy((8f * paneScale).dp)) {
-                        PaneSectionLabel(stringResource(R.string.input_controls_select_label_theme))
-                        InputControlsSimpleDropdown(
-                            options = state.inputControlsLabelThemeNames,
-                            selectedIndex = state.inputControlsSelectedLabelThemeIndex,
-                            onSelected = listener::onInputControlsLabelThemeSelected,
-                        )
-                    }
-                }
-
-                NavBooleanRow(
-                    title = stringResource(R.string.session_drawer_show_touchscreen_controls),
-                    checked = state.inputControlsShowOverlay,
-                    onCheckedChange = listener::onInputControlsShowOverlayChanged,
-                )
-
-                if (state.inputControlsShowOverlay) {
-                    NavSliderRow(
-                        label = stringResource(R.string.input_controls_editor_overlay_opacity),
-                        valueText = "${(state.inputControlsOverlayOpacity * 100).toInt()}%",
-                        value = state.inputControlsOverlayOpacity,
-                        valueRange = 0.1f..1.0f,
-                        steps = 8,
-                        onValueChange = listener::onInputControlsOverlayOpacityChanged,
-                    )
-                    Spacer(Modifier.height(4.dp))
-
-                    NavBooleanRow(
-                        title = stringResource(R.string.input_controls_tap_to_click),
-                        checked = state.inputControlsTapToClick,
-                        onCheckedChange = listener::onInputControlsTapToClickChanged,
-                    )
-                }
-
-                NavBooleanRow(
-                    title = stringResource(R.string.settings_general_touchscreen_haptics),
-                    checked = state.inputControlsTouchscreenHaptics,
-                    onCheckedChange = listener::onInputControlsTouchscreenHapticsChanged,
-                )
-
-                NavBooleanRow(
-                    title = stringResource(R.string.session_gamepad_enable_vibration),
-                    checked = state.inputControlsGamepadVibration,
-                    onCheckedChange = listener::onInputControlsGamepadVibrationChanged,
-                )
-
-                NavSliderRow(
-                    label = "Mouse sensitivity scale",
-                    valueText = "${Math.round(state.cursorSpeed * 100)}%",
-                    value = state.cursorSpeed * 100f,
-                    valueRange = 10f..300f,
-                    steps = 0,
-                    onValueChange = { listener.onCursorSpeedChanged(it / 100f) },
-                    adjustStep = 5f,
-                )
-
-                val rsMapMode = state.screenTouchMode == 2
-                val rsValue = if (rsMapMode) state.screenTouchRsSensitivity else state.rightStickSensitivity
-                NavSliderRow(
-                    label = stringResource(R.string.session_drawer_right_stick_sensitivity),
-                    valueText = "${Math.round(rsValue * 100)}%",
-                    value = rsValue * 100f,
-                    valueRange = (if (rsMapMode) 25f else 10f)..200f,
-                    steps = 0,
-                    onValueChange = { listener.onRightStickSensitivityChanged(it / 100f) },
-                    adjustStep = 5f,
-                )
-
-                LaunchedEffect(gcmEnabled) {
-                    if (gcmEnabled) scrollState.animateScrollTo(Int.MAX_VALUE)
-                }
-
-                NavBooleanRow(
-                    title = "GameSir Controller Rumble",
-                    subtitle = "For Android-mode GameSir controllers only",
-                    checked = gcmEnabled,
-                    onCheckedChange = { enabled ->
-                        listener.onInputControlsGcmRumbleModeChanged(if (enabled) "known" else "disabled")
-                    },
-                )
-
-                if (gcmEnabled) {
-                    Column(verticalArrangement = Arrangement.spacedBy((8f * paneScale).dp)) {
-                        Row(
-                            modifier = Modifier.fillMaxWidth(),
-                            horizontalArrangement = Arrangement.spacedBy((8f * paneScale).dp),
-                        ) {
-                            HUDToggleChip(
-                                label = "Known",
-                                checked = state.inputControlsGcmRumbleMode == "known",
-                                onClick = { listener.onInputControlsGcmRumbleModeChanged("known") },
-                                modifier = Modifier.weight(1f).paneNavItem(
-                                    cornerRadius = (16f * paneScale).dp,
-                                    onActivate = { listener.onInputControlsGcmRumbleModeChanged("known") },
-                                ),
-                            )
-                            HUDToggleChip(
-                                label = "All (experimental)",
-                                checked = state.inputControlsGcmRumbleMode == "all",
-                                onClick = { listener.onInputControlsGcmRumbleModeChanged("all") },
-                                modifier = Modifier.weight(1f).paneNavItem(
-                                    cornerRadius = (16f * paneScale).dp,
-                                    onActivate = { listener.onInputControlsGcmRumbleModeChanged("all") },
-                                ),
-                            )
-                        }
-                        Text(
-                            text = if (state.inputControlsGcmRumbleMode == "all")
-                                "All GameSir devices"
-                            else
-                                "G8+ MFi, X5s, X3 Pro",
-                            color = DrawerTextSecondary,
-                            fontSize = (11f * paneScale).sp,
-                        )
-                    }
-                }
-            }
-        }
-    }
-}
-
-/** Compact dropdown for the Controls Style/Label-Theme rows; like [InputControlsProfileSelector] but without the edit button (built-in, non-editable choices). */
-@Composable
-private fun InputControlsSimpleDropdown(
-    options: List<String>,
-    selectedIndex: Int,
-    onSelected: (Int) -> Unit,
-) {
-    val paneScale = LocalPaneScale.current
-    var expanded by remember { mutableStateOf(false) }
-    val selectedText = options.getOrElse(selectedIndex) { options.firstOrNull() ?: "" }
-    val parentNav = LocalPaneNav.current
-    val optionRegistry = remember { PaneNavRegistry() }
-    LaunchedEffect(expanded) {
-        if (expanded) {
-            optionRegistry.reset()
-            optionRegistry.controllerActive = true
-            parentNav?.overlay = optionRegistry
-            parentNav?.overlayClose = { expanded = false }
-        } else if (parentNav?.overlay === optionRegistry) {
-            parentNav.overlay = null
-            parentNav.overlayClose = null
-        }
-    }
-
-    val cornerRadius = (14f * paneScale).dp
-    val shape = RoundedCornerShape(cornerRadius)
-    val interactionSource = remember { MutableInteractionSource() }
-    val pressed = interactionSource.collectIsPressedAsState().value
-    val bgColor by animateColorAsState(
-        targetValue = if (pressed) PaneInnerPressed else PaneInnerResting,
-        animationSpec = tween(140),
-        label = "inputControlsSimpleBg",
-    )
-
-    Box(modifier = Modifier.fillMaxWidth()) {
-        Row(
-            modifier =
-                Modifier
-                    .fillMaxWidth()
-                    .clip(shape)
-                    .background(bgColor)
-                    .border(1.dp, RestingCardBorder, shape)
-                    .paneNavItem(
-                        cornerRadius = cornerRadius,
-                        onActivate = { expanded = true },
-                        onAdjust = { dir ->
-                            if (options.isNotEmpty()) {
-                                onSelected(((selectedIndex + dir) % options.size + options.size) % options.size)
-                            }
-                        },
-                    )
-                    .clickable(
-                        interactionSource = interactionSource,
-                        indication = null,
-                    ) { expanded = true }
-                    .padding(horizontal = (12f * paneScale).dp, vertical = (10f * paneScale).dp),
-            verticalAlignment = Alignment.CenterVertically,
-        ) {
-            Text(
-                text = selectedText,
-                color = DrawerTextPrimary,
-                fontSize = (14f * paneScale).sp,
-                fontWeight = FontWeight.Medium,
-                maxLines = 1,
-                overflow = TextOverflow.Ellipsis,
-                modifier = Modifier.weight(1f),
-            )
-            Icon(
-                imageVector = Icons.Outlined.ArrowDropDown,
-                contentDescription = null,
-                tint = DrawerTextSecondary,
-                modifier = Modifier.size((22f * paneScale).dp),
-            )
-        }
-
-        InputControlsOptionsPopup(
-            expanded = expanded,
-            options = options,
-            selectedIndex = selectedIndex,
-            onSelected = onSelected,
-            onDismiss = { expanded = false },
-            optionRegistry = optionRegistry,
-        )
-    }
-}
-
-@Composable
-private fun InputControlsProfileSelector(
-    profileNames: List<String>,
-    selectedIndex: Int,
-    onProfileSelected: (Int) -> Unit,
-    onEditClick: () -> Unit,
-) {
-    val paneScale = LocalPaneScale.current
-    var expanded by remember { mutableStateOf(false) }
-    val disabledPlaceholder = stringResource(R.string.common_ui_disabled_placeholder)
-    val selectedText = profileNames.getOrElse(selectedIndex) { disabledPlaceholder }
-    val parentNav = LocalPaneNav.current
-    val optionRegistry = remember { PaneNavRegistry() }
-    LaunchedEffect(expanded) {
-        if (expanded) {
-            optionRegistry.reset()
-            optionRegistry.controllerActive = true
-            parentNav?.overlay = optionRegistry
-            parentNav?.overlayClose = { expanded = false }
-        } else if (parentNav?.overlay === optionRegistry) {
-            parentNav.overlay = null
-            parentNav.overlayClose = null
-        }
-    }
-
-    val cornerRadius = (14f * paneScale).dp
-    val shape = RoundedCornerShape(cornerRadius)
-    val interactionSource = remember { MutableInteractionSource() }
-    val pressed = interactionSource.collectIsPressedAsState().value
-    val bgColor by animateColorAsState(
-        targetValue = if (pressed) PaneInnerPressed else PaneInnerResting,
-        animationSpec = tween(140),
-        label = "inputControlsProfileBg",
-    )
-
-    Row(
-        modifier = Modifier.fillMaxWidth(),
-        horizontalArrangement = Arrangement.spacedBy((8f * paneScale).dp),
-        verticalAlignment = Alignment.CenterVertically,
-    ) {
-        Box(modifier = Modifier.weight(1f)) {
-            Row(
-                modifier =
-                    Modifier
-                        .fillMaxWidth()
-                        .clip(shape)
-                        .background(bgColor)
-                        .border(1.dp, RestingCardBorder, shape)
-                        .paneNavItem(cornerRadius = cornerRadius, onActivate = { expanded = true })
-                        .clickable(
-                            interactionSource = interactionSource,
-                            indication = null,
-                        ) { expanded = true }
-                        .padding(horizontal = (12f * paneScale).dp, vertical = (10f * paneScale).dp),
-                verticalAlignment = Alignment.CenterVertically,
-            ) {
-                Text(
-                    text = selectedText,
-                    color = DrawerTextPrimary,
-                    fontSize = (14f * paneScale).sp,
-                    fontWeight = FontWeight.Medium,
-                    maxLines = 1,
-                    overflow = TextOverflow.Ellipsis,
-                    modifier = Modifier.weight(1f),
-                )
-                Icon(
-                    imageVector = Icons.Outlined.ArrowDropDown,
-                    contentDescription = null,
-                    tint = DrawerTextSecondary,
-                    modifier = Modifier.size((22f * paneScale).dp),
-                )
-            }
-
-            InputControlsOptionsPopup(
-                expanded = expanded,
-                options = profileNames,
-                selectedIndex = selectedIndex,
-                onSelected = onProfileSelected,
-                onDismiss = { expanded = false },
-                optionRegistry = optionRegistry,
-            )
-        }
-
-        Box(
-            modifier =
-                Modifier
-                    .size((44f * paneScale).dp)
-                    .clip(shape)
-                    .background(PaneInnerResting)
-                    .border(1.dp, RestingCardBorder, shape)
-                    .paneNavItem(cornerRadius = cornerRadius, onActivate = onEditClick)
-                    .clickable(onClick = onEditClick),
-            contentAlignment = Alignment.Center,
-        ) {
-            Icon(
-                imageVector = Icons.Outlined.Settings,
-                contentDescription = stringResource(R.string.common_ui_settings),
-                tint = DrawerTextPrimary,
-                modifier = Modifier.size((20f * paneScale).dp),
-            )
-        }
-    }
-}
-
-// Drawer-styled dropdown for the Controls selectors.
-@Composable
-private fun InputControlsOptionsPopup(
-    expanded: Boolean,
-    options: List<String>,
-    selectedIndex: Int,
-    onSelected: (Int) -> Unit,
-    onDismiss: () -> Unit,
-    optionRegistry: PaneNavRegistry,
-) {
-    if (!expanded) return
-    val paneScale = LocalPaneScale.current
-    val density = LocalDensity.current
-    val gapPx = with(density) { (4f * paneScale).dp.roundToPx() }
-    val shape = RoundedCornerShape((12f * paneScale).dp)
-    val scrollState = rememberScrollState()
-    Popup(
-        popupPositionProvider = remember(gapPx) { TaskManagerPopupPositionProvider(gapPx) },
-        onDismissRequest = onDismiss,
-        properties = PopupProperties(focusable = false),
-    ) {
-        CompositionLocalProvider(LocalPaneNav provides optionRegistry) {
-            Column(
-                modifier =
-                    Modifier
-                        .widthIn(min = (160f * paneScale).dp, max = (280f * paneScale).dp)
-                        .clip(shape)
-                        .background(PaneSurfaceColor)
-                        .border(1.dp, RestingCardBorder, shape)
-                        .heightIn(max = (260f * paneScale).dp)
-                        .verticalScroll(scrollState)
-                        .padding((5f * paneScale).dp),
-                verticalArrangement = Arrangement.spacedBy((4f * paneScale).dp),
-            ) {
-                options.forEachIndexed { index, name ->
-                    InputControlsOptionItem(
-                        label = name,
-                        selected = index == selectedIndex,
-                        onClick = {
-                            onSelected(index)
-                            onDismiss()
-                        },
-                    )
-                }
-            }
-        }
-    }
-}
-
-@Composable
-private fun InputControlsOptionItem(
-    label: String,
-    selected: Boolean,
-    onClick: () -> Unit,
-    modifier: Modifier = Modifier,
-) {
-    val paneScale = LocalPaneScale.current
-    val interactionSource = remember { MutableInteractionSource() }
-    val pressed = interactionSource.collectIsPressedAsState().value
-    val bgColor by animateColorAsState(
-        targetValue = if (pressed) DrawerAccent.copy(alpha = 0.16f) else PaneInnerResting,
-        animationSpec = tween(120),
-        label = "inputControlsOptionItem",
-    )
-    val shape = RoundedCornerShape((8f * paneScale).dp)
-    Row(
-        modifier =
-            modifier
-                .fillMaxWidth()
-                .clip(shape)
-                .background(bgColor)
-                .border(1.dp, if (selected) ActiveCardBorder else RestingCardBorder, shape)
-                .paneNavItem(cornerRadius = (8f * paneScale).dp, onActivate = onClick)
-                .clickable(interactionSource = interactionSource, indication = null, onClick = onClick)
-                .padding(horizontal = (12f * paneScale).dp, vertical = (10f * paneScale).dp),
-        verticalAlignment = Alignment.CenterVertically,
-        horizontalArrangement = Arrangement.spacedBy((8f * paneScale).dp),
-    ) {
-        Text(
-            text = label,
-            color = if (selected) DrawerAccent else DrawerTextPrimary,
-            fontSize = (13f * paneScale).sp,
-            fontWeight = if (selected) FontWeight.SemiBold else FontWeight.Medium,
-            maxLines = 1,
-            overflow = TextOverflow.Ellipsis,
-            modifier = Modifier.weight(1f),
-        )
-        if (selected) {
-            Icon(
-                imageVector = Icons.Outlined.Check,
-                contentDescription = null,
-                tint = DrawerAccent,
-                modifier = Modifier.size((16f * paneScale).dp),
-            )
-        }
-    }
-}
-
-@Composable
-private fun ExpandableSection(
-    title: String,
-    expanded: Boolean,
-    onToggle: () -> Unit,
-    content: @Composable () -> Unit,
-) {
-    val paneScale = LocalPaneScale.current
-    val rotation by animateFloatAsState(
-        targetValue = if (expanded) 180f else 0f,
-        animationSpec = tween(180, easing = FastOutSlowInEasing),
-        label = "expandableRotation",
-    )
-    val headerInteractionSource = remember { MutableInteractionSource() }
-    val headerPressed = headerInteractionSource.collectIsPressedAsState().value
-    val headerBg by animateColorAsState(
-        targetValue =
-            when {
-                headerPressed -> PaneInnerPressed
-                else -> PaneInnerResting
-            },
-        animationSpec = tween(140),
-        label = "expandableHeaderBg",
-    )
-    val headerBorder by animateColorAsState(
-        targetValue = if (expanded) DrawerAccent else RestingCardBorder,
-        animationSpec = tween(140),
-        label = "expandableHeaderBorder",
-    )
-    val headerShape = RoundedCornerShape((12f * paneScale).dp)
-    Column(verticalArrangement = Arrangement.spacedBy((10f * paneScale).dp)) {
-        Row(
-            modifier =
-                Modifier
-                    .fillMaxWidth()
-                    .clip(headerShape)
-                    .background(headerBg)
-                    .border(1.dp, headerBorder, headerShape)
-                    .paneNavItem(cornerRadius = (12f * paneScale).dp, onActivate = onToggle)
-                    .clickable(
-                        interactionSource = headerInteractionSource,
-                        indication = null,
-                        onClick = onToggle,
-                    )
-                    .padding(horizontal = (12f * paneScale).dp, vertical = (10f * paneScale).dp),
-            verticalAlignment = Alignment.CenterVertically,
-        ) {
-            Text(
-                text = title,
-                color = if (expanded) DrawerAccent else DrawerTextPrimary,
-                fontSize = (14f * paneScale).sp,
-                fontWeight = FontWeight.SemiBold,
-                letterSpacing = 0.3.sp,
-                modifier = Modifier.weight(1f),
-            )
-            Icon(
-                imageVector = Icons.Outlined.ExpandMore,
-                contentDescription = null,
-                tint = if (expanded) DrawerAccent else DrawerTextSecondary,
-                modifier =
-                    Modifier
-                        .size((18f * paneScale).dp)
-                        .graphicsLayer { rotationZ = rotation },
-            )
-        }
-        AnimatedVisibility(visible = expanded) {
-            Column(verticalArrangement = Arrangement.spacedBy((12f * paneScale).dp)) {
-                content()
-            }
-        }
-    }
-}
-
-@Composable
-private fun ScreenEffectsPaneContent(
-    state: XServerDrawerState,
-    listener: XServerDrawerActionListener,
-) {
-    BoxWithConstraints(modifier = Modifier.fillMaxSize()) {
-        val paneScale = computePaneScale(maxHeight)
-        CompositionLocalProvider(LocalPaneScale provides paneScale) {
-            Column(
-                modifier =
-                    Modifier
-                        .fillMaxWidth()
-                        .verticalScroll(rememberScrollState())
-                        .padding(horizontal = (12f * paneScale).dp, vertical = (12f * paneScale).dp),
-                verticalArrangement = Arrangement.spacedBy((10f * paneScale).dp),
-            ) {
-                Column(verticalArrangement = Arrangement.spacedBy((8f * paneScale).dp)) {
-                    PaneSectionLabel(stringResource(R.string.shortcuts_graphics_sgsr_full_title))
-                    NavBooleanRow(
-                        title = stringResource(R.string.session_drawer_upscaler_fsr),
-                        checked = state.sgsrEnabled,
-                        onCheckedChange = listener::onSGSREnabledChanged,
-                    )
-
-                    AnimatedVisibility(
-                        visible = state.sgsrEnabled,
-                        enter =
-                            expandVertically(
-                                animationSpec = tween(durationMillis = 220, easing = FastOutSlowInEasing),
-                                expandFrom = Alignment.Top,
-                            ) + fadeIn(animationSpec = tween(durationMillis = 160, easing = FastOutSlowInEasing)),
-                        exit =
-                            shrinkVertically(
-                                animationSpec = tween(durationMillis = 180, easing = FastOutSlowInEasing),
-                                shrinkTowards = Alignment.Top,
-                            ) + fadeOut(animationSpec = tween(durationMillis = 120, easing = FastOutSlowInEasing)),
-                    ) {
-                        Column(verticalArrangement = Arrangement.spacedBy((8f * paneScale).dp)) {
-                            NavSliderRow(
-                                label = stringResource(R.string.session_drawer_sgsr_edge_sharpness),
-                                valueText = "${state.sgsrSharpness}%",
-                                value = state.sgsrSharpness.toFloat(),
-                                valueRange = 0f..100f,
-                                steps = 99,
-                                onValueChange = { listener.onSGSRSharpnessChanged(it.roundToInt().coerceIn(0, 100)) },
-                            )
-                        }
-                    }
-                }
-
-                ThinDivider()
-
-                Column(verticalArrangement = Arrangement.spacedBy((8f * paneScale).dp)) {
-                    PaneSectionLabel(stringResource(R.string.session_drawer_color_profile))
-
-                    val profiles =
-                        listOf(
-                            0 to stringResource(R.string.session_drawer_color_profile_disabled),
-                            1 to stringResource(R.string.session_drawer_color_profile_hdr),
-                            2 to stringResource(R.string.session_drawer_color_profile_natural),
-                            4 to stringResource(R.string.session_drawer_color_profile_toon),
-                            3 to stringResource(R.string.session_drawer_color_profile_crt),
-                            5 to stringResource(R.string.session_drawer_color_profile_ntsc),
-                            6 to stringResource(R.string.session_drawer_color_profile_ntsc2),
-                        )
-
-                    ChipFlow {
-                        profiles.forEach { (id, label) ->
-                            HUDToggleChip(
-                                label = label,
-                                checked = state.colorProfile == id,
-                                onClick = { listener.onColorProfileSelected(id) },
-                                modifier = Modifier.paneNavItem(
-                                    cornerRadius = (16f * paneScale).dp,
-                                    onActivate = { listener.onColorProfileSelected(id) },
-                                ),
-                            )
-                        }
-                    }
-
-                    NavBooleanRow(
-                        title = stringResource(R.string.session_drawer_vivid),
-                        checked = state.vividEnabled,
-                        onCheckedChange = listener::onVividEnabledChanged,
-                    )
-
-                    if (state.vividEnabled) {
-                        NavSliderRow(
-                            label = stringResource(R.string.session_drawer_vivid_strength),
-                            valueText = "${state.vividStrength}%",
-                            value = state.vividStrength.toFloat(),
-                            valueRange = 0f..100f,
-                            steps = 99,
-                            onValueChange = { listener.onVividStrengthChanged(it.roundToInt()) },
-                        )
-                    }
-
-                    NavBooleanRow(
-                        title = stringResource(R.string.session_drawer_sharpen),
-                        checked = state.sharpenEnabled,
-                        onCheckedChange = listener::onSharpenEnabledChanged,
-                    )
-
-                    if (state.sharpenEnabled) {
-                        NavSliderRow(
-                            label = stringResource(R.string.session_drawer_strength),
-                            valueText = "${state.sharpenStrength}%",
-                            value = state.sharpenStrength.toFloat(),
-                            valueRange = 0f..100f,
-                            steps = 99,
-                            onValueChange = { listener.onSharpenStrengthChanged(it.roundToInt().coerceIn(0, 100)) },
-                        )
-                    }
-
-                    NavBooleanRow(
-                        title = stringResource(R.string.session_drawer_scanlines),
-                        checked = state.scanlinesEnabled,
-                        onCheckedChange = listener::onScanlinesEnabledChanged,
-                    )
-
-                    if (state.scanlinesEnabled) {
-                        NavSliderRow(
-                            label = stringResource(R.string.session_drawer_intensity),
-                            valueText = "${state.scanlinesIntensity}%",
-                            value = state.scanlinesIntensity.toFloat(),
-                            valueRange = 0f..100f,
-                            steps = 99,
-                            onValueChange = { listener.onScanlinesIntensityChanged(it.roundToInt().coerceIn(0, 100)) },
-                        )
-                    }
-
-                    NavBooleanRow(
-                        title = stringResource(R.string.session_drawer_pixelate),
-                        checked = state.pixelateEnabled,
-                        onCheckedChange = listener::onPixelateEnabledChanged,
-                    )
-
-                    if (state.pixelateEnabled) {
-                        NavSliderRow(
-                            label = stringResource(R.string.session_drawer_block_size),
-                            valueText = "${state.pixelateBlock}px",
-                            value = state.pixelateBlock.toFloat(),
-                            valueRange = 2f..14f,
-                            steps = 11,
-                            onValueChange = { listener.onPixelateBlockChanged(it.roundToInt().coerceIn(2, 14)) },
-                        )
-                    }
-
-                    NavSliderRow(
-                        label = stringResource(R.string.session_drawer_brightness),
-                        valueText = "${state.brightness}",
-                        value = state.brightness.toFloat(),
-                        valueRange = -100f..100f,
-                        steps = 39,
-                        onValueChange = { listener.onBrightnessChanged(it.roundToInt().coerceIn(-100, 100)) },
-                    )
-
-                    NavSliderRow(
-                        label = stringResource(R.string.session_drawer_contrast),
-                        valueText = "${state.contrast}",
-                        value = state.contrast.toFloat(),
-                        valueRange = -100f..100f,
-                        steps = 39,
-                        onValueChange = { listener.onContrastChanged(it.roundToInt().coerceIn(-100, 100)) },
-                    )
-
-                    NavSliderRow(
-                        label = stringResource(R.string.session_drawer_gamma),
-                        valueText = String.format("%.2fx", state.gammaPercent / 100f),
-                        value = state.gammaPercent.toFloat(),
-                        valueRange = 50f..250f,
-                        steps = 19,
-                        onValueChange = { listener.onGammaChanged(it.roundToInt().coerceIn(50, 250)) },
-                    )
-
-                    NavSliderRow(
-                        label = stringResource(R.string.session_drawer_saturation),
-                        valueText = "${state.saturation}%",
-                        value = state.saturation.toFloat(),
-                        valueRange = 0f..200f,
-                        steps = 39,
-                        onValueChange = { listener.onSaturationChanged(it.roundToInt().coerceIn(0, 200)) },
-                    )
-
-                    NavSliderRow(
-                        label = stringResource(R.string.session_drawer_temperature),
-                        valueText = "${state.temperature}",
-                        value = state.temperature.toFloat(),
-                        valueRange = -100f..100f,
-                        steps = 39,
-                        onValueChange = { listener.onTemperatureChanged(it.roundToInt().coerceIn(-100, 100)) },
-                    )
-
-                    NavSliderRow(
-                        label = stringResource(R.string.session_drawer_tint),
-                        valueText = "${state.tint}",
-                        value = state.tint.toFloat(),
-                        valueRange = -100f..100f,
-                        steps = 39,
-                        onValueChange = { listener.onTintChanged(it.roundToInt().coerceIn(-100, 100)) },
-                    )
-
-                    Box(
-                        Modifier.fillMaxWidth().paneNavItem(
-                            cornerRadius = (12f * paneScale).dp,
-                            onActivate = { listener.onResetEffects() },
-                        ),
-                    ) {
-                        DrawerResetRow(
-                            label = stringResource(R.string.session_drawer_reset_effects),
-                            onClick = listener::onResetEffects,
-                        )
-                    }
-                }
-
-                ThinDivider()
-
-                Column(verticalArrangement = Arrangement.spacedBy((8f * paneScale).dp)) {
-                    PaneSectionLabel(stringResource(R.string.session_drawer_color_blind))
-
-                    Row(horizontalArrangement = Arrangement.spacedBy((8f * paneScale).dp)) {
-                        HUDToggleChip(
-                            label = stringResource(R.string.session_drawer_color_blind_protan),
-                            checked = state.colorBlind == 1,
-                            onClick = { listener.onColorBlindSelected(if (state.colorBlind == 1) 0 else 1) },
-                            modifier = Modifier.weight(1f).paneNavItem(
-                                cornerRadius = (16f * paneScale).dp,
-                                onActivate = { listener.onColorBlindSelected(if (state.colorBlind == 1) 0 else 1) },
-                            ),
-                        )
-                        HUDToggleChip(
-                            label = stringResource(R.string.session_drawer_color_blind_deutan),
-                            checked = state.colorBlind == 2,
-                            onClick = { listener.onColorBlindSelected(if (state.colorBlind == 2) 0 else 2) },
-                            modifier = Modifier.weight(1f).paneNavItem(
-                                cornerRadius = (16f * paneScale).dp,
-                                onActivate = { listener.onColorBlindSelected(if (state.colorBlind == 2) 0 else 2) },
-                            ),
-                        )
-                        HUDToggleChip(
-                            label = stringResource(R.string.session_drawer_color_blind_tritan),
-                            checked = state.colorBlind == 3,
-                            onClick = { listener.onColorBlindSelected(if (state.colorBlind == 3) 0 else 3) },
-                            modifier = Modifier.weight(1f).paneNavItem(
-                                cornerRadius = (16f * paneScale).dp,
-                                onActivate = { listener.onColorBlindSelected(if (state.colorBlind == 3) 0 else 3) },
-                            ),
-                        )
-                    }
-                }
-
-                ThinDivider()
-
-                Column(verticalArrangement = Arrangement.spacedBy((8f * paneScale).dp)) {
-                    PaneSectionLabel(stringResource(R.string.session_drawer_scale))
-
-                    NavBooleanRow(
-                        title = stringResource(R.string.session_drawer_scale_nearest),
-                        checked = state.scaleFilter == 1,
-                        onCheckedChange = { on -> listener.onScaleFilterSelected(if (on) 1 else 0) },
-                    )
-
-                    NavBooleanRow(
-                        title = stringResource(R.string.session_drawer_scale_linear),
-                        checked = state.scaleFilter == 2,
-                        onCheckedChange = { on -> listener.onScaleFilterSelected(if (on) 2 else 0) },
-                    )
-
-                    NavBooleanRow(
-                        title = stringResource(R.string.session_drawer_scale_bicubic),
-                        checked = state.scaleFilter == 3,
-                        onCheckedChange = { on -> listener.onScaleFilterSelected(if (on) 3 else 0) },
-                    )
-                }
-            }
-        }
-    }
-}
-
-
-@Composable
-private fun OutputPaneContent(
-    state: XServerDrawerState,
-    listener: XServerDrawerActionListener,
-) {
-    BoxWithConstraints(modifier = Modifier.fillMaxSize()) {
-        val paneScale = computePaneScale(maxHeight)
-        CompositionLocalProvider(LocalPaneScale provides paneScale) {
-            Column(
-                modifier =
-                    Modifier
-                        .fillMaxWidth()
-                        .verticalScroll(rememberScrollState())
-                        .padding(horizontal = (12f * paneScale).dp, vertical = (12f * paneScale).dp),
-                verticalArrangement = Arrangement.spacedBy((10f * paneScale).dp),
-            ) {
-                if (state.outputSwapActive) {
-                    OutputActiveControls(state = state, listener = listener, paneScale = paneScale)
-                } else if (state.outputDisplayAvailable) {
-                    OutputSendToDisplay(state = state, listener = listener, paneScale = paneScale)
-                } else {
-                    OutputCastEntry(listener = listener, paneScale = paneScale)
-                }
-            }
-        }
-    }
-}
-
-@Composable
-private fun OutputActiveControls(
-    state: XServerDrawerState,
-    listener: XServerDrawerActionListener,
-    paneScale: Float,
-) {
-    OutputDeviceHeader(state = state, paneScale = paneScale)
-
-    OutputCard(paneScale = paneScale, title = stringResource(R.string.session_drawer_output_display)) {
-        if (state.outputResolutionLabels.isNotEmpty()) {
-            Column(verticalArrangement = Arrangement.spacedBy((6f * paneScale).dp)) {
-                OutputFieldLabel(stringResource(R.string.session_drawer_output_resolution), paneScale)
-                InputControlsSimpleDropdown(
-                    options = state.outputResolutionLabels,
-                    selectedIndex = state.outputSelectedResolutionIndex,
-                    onSelected = listener::onOutputResolutionSelected,
-                )
-                Text(
-                    text = if (state.outputPanelScaling) {
-                        stringResource(R.string.session_drawer_output_scaling_note, state.outputPanelNative)
-                    } else {
-                        stringResource(R.string.session_drawer_output_render_note)
-                    },
-                    color = DrawerTextSecondary,
-                    fontSize = (11f * paneScale).sp,
-                    lineHeight = (15f * paneScale).sp,
-                )
-            }
-        }
-        if (!state.outputPanelScaling && state.outputRefreshLabels.isNotEmpty()) {
-            Column(verticalArrangement = Arrangement.spacedBy((6f * paneScale).dp)) {
-                OutputFieldLabel(stringResource(R.string.session_drawer_output_refresh_rate), paneScale)
-                InputControlsSimpleDropdown(
-                    options = state.outputRefreshLabels,
-                    selectedIndex = state.outputSelectedRefreshIndex,
-                    onSelected = listener::onOutputRefreshRateSelected,
-                )
-            }
-        }
-        Column(verticalArrangement = Arrangement.spacedBy((6f * paneScale).dp)) {
-            OutputFieldLabel(stringResource(R.string.session_drawer_output_aspect_ratio), paneScale)
-            val aspectLabels =
-                listOf(
-                    stringResource(R.string.session_drawer_output_aspect_fit),
-                    stringResource(R.string.session_drawer_output_aspect_stretch),
-                    stringResource(R.string.session_drawer_output_aspect_zoom),
-                )
-            ChipFlow {
-                aspectLabels.forEachIndexed { index, label ->
-                    HUDToggleChip(
-                        label = label,
-                        checked = state.outputAspectMode == index,
-                        onClick = { listener.onOutputAspectModeSelected(index) },
-                        modifier = Modifier.paneNavItem(
-                            cornerRadius = (12f * paneScale).dp,
-                            onActivate = { listener.onOutputAspectModeSelected(index) },
-                        ),
-                    )
-                }
-            }
-        }
-    }
-
-    if (state.outputGameModeSupported) {
-        OutputCard(paneScale = paneScale, title = stringResource(R.string.session_drawer_output_game_mode)) {
-            ChipFlow {
-                HUDToggleChip(
-                    label = stringResource(R.string.session_drawer_output_game_mode_on),
-                    checked = state.outputGameModeEnabled,
-                    onClick = { listener.onOutputGameModeToggled(true) },
-                    modifier = Modifier.paneNavItem(
-                        cornerRadius = (12f * paneScale).dp,
-                        onActivate = { listener.onOutputGameModeToggled(true) },
-                    ),
-                )
-                HUDToggleChip(
-                    label = stringResource(R.string.session_drawer_output_game_mode_off),
-                    checked = !state.outputGameModeEnabled,
-                    onClick = { listener.onOutputGameModeToggled(false) },
-                    modifier = Modifier.paneNavItem(
-                        cornerRadius = (12f * paneScale).dp,
-                        onActivate = { listener.onOutputGameModeToggled(false) },
-                    ),
-                )
-            }
-            Text(
-                text = stringResource(R.string.session_drawer_output_game_mode_note),
-                color = DrawerTextSecondary,
-                fontSize = (11f * paneScale).sp,
-                lineHeight = (15f * paneScale).sp,
-            )
-        }
-    }
-
-    if (state.outputVitureConnected) {
-        OutputGlassesCard(state = state, listener = listener, paneScale = paneScale)
-    }
-
-    OutputPaneButton(
-        label = stringResource(R.string.session_drawer_output_return_to_phone),
-        paneScale = paneScale,
-        onClick = listener::onOutputReturnToPhone,
-    )
-}
-
-@Composable
-private fun OutputGlassesCard(
-    state: XServerDrawerState,
-    listener: XServerDrawerActionListener,
-    paneScale: Float,
-) {
-    OutputCard(
-        paneScale = paneScale,
-        title = state.outputVitureName.ifEmpty { stringResource(R.string.session_drawer_output_glasses) },
-    ) {
-        if (state.outputVitureSupportsBrightness) {
-            DrawerSliderRow(
-                label = stringResource(R.string.session_drawer_output_brightness),
-                valueText = "${state.outputVitureBrightness}/${state.outputVitureBrightnessMax}",
-                value = state.outputVitureBrightness.toFloat(),
-                valueRange = 0f..state.outputVitureBrightnessMax.toFloat(),
-                steps = (state.outputVitureBrightnessMax - 1).coerceAtLeast(0),
-                onValueChange = { listener.onOutputVitureBrightness(it.roundToInt()) },
-            )
-        }
-        if (state.outputVitureSupportsFilm) {
-            if (state.outputVitureFilmStepped) {
-                DrawerSliderRow(
-                    label = stringResource(R.string.session_drawer_output_shade),
-                    valueText = "${state.outputVitureFilm}/8",
-                    value = state.outputVitureFilm.toFloat(),
-                    valueRange = 0f..8f,
-                    steps = 7,
-                    onValueChange = { listener.onOutputVitureFilm(it.roundToInt()) },
-                )
-            } else {
-                Column(verticalArrangement = Arrangement.spacedBy((6f * paneScale).dp)) {
-                    OutputFieldLabel(stringResource(R.string.session_drawer_output_shade), paneScale)
-                    ChipFlow {
-                        HUDToggleChip(
-                            label = stringResource(R.string.session_drawer_output_game_mode_on),
-                            checked = state.outputVitureFilm > 0,
-                            onClick = { listener.onOutputVitureFilm(1) },
-                            modifier = Modifier.paneNavItem(
-                                cornerRadius = (12f * paneScale).dp,
-                                onActivate = { listener.onOutputVitureFilm(1) },
-                            ),
-                        )
-                        HUDToggleChip(
-                            label = stringResource(R.string.session_drawer_output_game_mode_off),
-                            checked = state.outputVitureFilm == 0,
-                            onClick = { listener.onOutputVitureFilm(0) },
-                            modifier = Modifier.paneNavItem(
-                                cornerRadius = (12f * paneScale).dp,
-                                onActivate = { listener.onOutputVitureFilm(0) },
-                            ),
-                        )
-                    }
-                }
-            }
-        }
-        if (state.outputVitureSupportsVolume) {
-            DrawerSliderRow(
-                label = stringResource(R.string.session_drawer_output_volume),
-                valueText = "${state.outputVitureVolume}/${state.outputVitureVolumeMax}",
-                value = state.outputVitureVolume.toFloat(),
-                valueRange = 0f..state.outputVitureVolumeMax.toFloat(),
-                steps = (state.outputVitureVolumeMax - 1).coerceAtLeast(0),
-                onValueChange = { listener.onOutputVitureVolume(it.roundToInt()) },
-            )
-        }
-        if (state.outputVitureSupports3D) {
-            Column(verticalArrangement = Arrangement.spacedBy((6f * paneScale).dp)) {
-                OutputFieldLabel(stringResource(R.string.session_drawer_output_3d), paneScale)
-                ChipFlow {
-                    HUDToggleChip(
-                        label = stringResource(R.string.session_drawer_output_game_mode_on),
-                        checked = state.outputViture3D,
-                        onClick = { listener.onOutputViture3D(true) },
-                        modifier = Modifier.paneNavItem(
-                            cornerRadius = (12f * paneScale).dp,
-                            onActivate = { listener.onOutputViture3D(true) },
-                        ),
-                    )
-                    HUDToggleChip(
-                        label = stringResource(R.string.session_drawer_output_game_mode_off),
-                        checked = !state.outputViture3D,
-                        onClick = { listener.onOutputViture3D(false) },
-                        modifier = Modifier.paneNavItem(
-                            cornerRadius = (12f * paneScale).dp,
-                            onActivate = { listener.onOutputViture3D(false) },
-                        ),
-                    )
-                }
-            }
-        }
-        Text(
-            text = stringResource(R.string.session_drawer_output_glasses_note),
-            color = DrawerTextSecondary,
-            fontSize = (11f * paneScale).sp,
-            lineHeight = (15f * paneScale).sp,
-        )
-    }
-}
-
-@Composable
-private fun OutputCard(
-    paneScale: Float,
-    title: String,
-    content: @Composable () -> Unit,
-) {
-    Column(
-        modifier =
-            Modifier
-                .fillMaxWidth()
-                .clip(RoundedCornerShape((14f * paneScale).dp))
-                .background(PaneInnerResting)
-                .border(1.dp, RestingCardBorder, RoundedCornerShape((14f * paneScale).dp))
-                .padding(horizontal = (12f * paneScale).dp, vertical = (12f * paneScale).dp),
-        verticalArrangement = Arrangement.spacedBy((10f * paneScale).dp),
-    ) {
-        PaneSectionLabel(title)
-        content()
-    }
-}
-
-@Composable
-private fun OutputFieldLabel(text: String, paneScale: Float) {
-    Text(
-        text = text,
-        color = DrawerTextSecondary,
-        fontSize = (12f * paneScale).sp,
-        fontWeight = FontWeight.Medium,
-    )
-}
-
-@Composable
-private fun OutputDeviceHeader(state: XServerDrawerState, paneScale: Float) {
-    Row(
-        verticalAlignment = Alignment.CenterVertically,
-        horizontalArrangement = Arrangement.spacedBy((8f * paneScale).dp),
-        modifier = Modifier.padding(horizontal = (2f * paneScale).dp),
-    ) {
-        Icon(
-            imageVector = Icons.Outlined.Monitor,
-            contentDescription = null,
-            tint = DrawerAccent,
-            modifier = Modifier.size((22f * paneScale).dp),
-        )
-        Text(
-            text = state.outputDisplayName.ifEmpty { stringResource(R.string.session_drawer_output_title) },
-            color = DrawerTextPrimary,
-            fontSize = (15f * paneScale).sp,
-            fontWeight = FontWeight.SemiBold,
-            maxLines = 1,
-            overflow = TextOverflow.Ellipsis,
-        )
-    }
-}
-
-@Composable
-private fun OutputSendToDisplay(
-    state: XServerDrawerState,
-    listener: XServerDrawerActionListener,
-    paneScale: Float,
-) {
-    OutputDeviceHeader(state = state, paneScale = paneScale)
-    OutputPaneButton(
-        label = stringResource(R.string.session_drawer_output_send_to_display),
-        paneScale = paneScale,
-        onClick = listener::onOutputSwapToDisplay,
-    )
-    Text(
-        text = stringResource(R.string.session_drawer_output_send_note),
-        color = DrawerTextSecondary,
-        fontSize = (11f * paneScale).sp,
-        lineHeight = (15f * paneScale).sp,
-    )
-}
-
-@Composable
-private fun OutputCastEntry(
-    listener: XServerDrawerActionListener,
-    paneScale: Float,
-) {
-    Column(verticalArrangement = Arrangement.spacedBy((6f * paneScale).dp)) {
-        PaneSectionLabel(stringResource(R.string.session_drawer_output_cast_title))
-        Text(
-            text = stringResource(R.string.session_drawer_output_cast_body),
-            color = DrawerTextSecondary,
-            fontSize = (12f * paneScale).sp,
-            lineHeight = (16f * paneScale).sp,
-        )
-    }
-    OutputPaneButton(
-        label = stringResource(R.string.session_drawer_output_cast_button),
-        paneScale = paneScale,
-        onClick = listener::onOutputCastClick,
-    )
-    Text(
-        text = stringResource(R.string.session_drawer_output_cast_note),
-        color = DrawerTextSecondary,
-        fontSize = (11f * paneScale).sp,
-        lineHeight = (15f * paneScale).sp,
-    )
-}
-
-@Composable
-private fun OutputPaneButton(
-    label: String,
-    paneScale: Float,
-    onClick: () -> Unit,
-) {
-    Row(
-        modifier =
-            Modifier
-                .fillMaxWidth()
-                .clip(RoundedCornerShape((14f * paneScale).dp))
-                .background(PaneInnerResting)
-                .border(1.dp, RestingCardBorder, RoundedCornerShape((14f * paneScale).dp))
-                .paneNavItem(cornerRadius = (14f * paneScale).dp, onActivate = onClick)
-                .clickable { onClick() }
-                .padding(horizontal = (12f * paneScale).dp, vertical = (12f * paneScale).dp),
-        verticalAlignment = Alignment.CenterVertically,
-        horizontalArrangement = Arrangement.Center,
-    ) {
-        Text(
-            text = label,
-            color = DrawerTextPrimary,
-            fontSize = (14f * paneScale).sp,
-            fontWeight = FontWeight.SemiBold,
-        )
-    }
-}
-
-@Composable
-private fun TaskManagerPaneContent(
-    taskManagerState: TaskManagerPaneState,
-    listener: XServerDrawerActionListener,
-    onClose: () -> Unit,
-) {
-    var showNewTaskDialog by remember { mutableStateOf(false) }
-    var processPendingEnd by remember { mutableStateOf<TaskManagerProcess?>(null) }
-    var expandedAffinityPid by remember { mutableStateOf<Int?>(null) }
-    val pendingAffinities = remember { mutableStateMapOf<Int, PendingTaskAffinity>() }
-
-    DisposableEffect(Unit) {
-        listener.onTaskManagerVisibilityChanged(true)
-        onDispose { listener.onTaskManagerVisibilityChanged(false) }
-    }
-
-    LaunchedEffect(taskManagerState.processes) {
-        val visibleProcessPids = taskManagerState.processes.map { it.pid }.toSet()
-        val now = System.currentTimeMillis()
-        pendingAffinities.keys.toList().forEach { pid ->
-            if (pid !in visibleProcessPids) pendingAffinities.remove(pid)
-        }
-        taskManagerState.processes.forEach { process ->
-            val pending = pendingAffinities[process.pid]
-            if (
-                pending != null &&
-                    (pending.affinityMask == process.affinityMask ||
-                        now - pending.requestedAtMillis > PendingTaskAffinityTimeoutMs)
-            ) {
-                pendingAffinities.remove(process.pid)
-            }
-        }
-        if (expandedAffinityPid != null && expandedAffinityPid !in visibleProcessPids) {
-            expandedAffinityPid = null
-        }
-    }
-
-    BoxWithConstraints(modifier = Modifier.fillMaxSize()) {
-        val paneScale = computePaneScale(maxHeight)
-        val affinityCoreCount =
-            if (taskManagerState.cpuCoreCount > 0) {
-                taskManagerState.cpuCoreCount
-            } else {
-                Runtime.getRuntime().availableProcessors()
-            }
-        CompositionLocalProvider(LocalPaneScale provides paneScale) {
-            Column(
-                modifier =
-                    Modifier
-                        .fillMaxSize()
-                        .padding(horizontal = (12f * paneScale).dp, vertical = (10f * paneScale).dp),
-                verticalArrangement = Arrangement.spacedBy((10f * paneScale).dp),
-            ) {
-                TaskManagerHeader(
-                    cpuPercent = taskManagerState.cpuPercent,
-                    cpuCoreCount = taskManagerState.cpuCoreCount,
-                    cpuCorePercents = taskManagerState.cpuCorePercents,
-                    memoryPercent = taskManagerState.memoryPercent,
-                    memoryDetail = taskManagerState.memoryDetail,
-                    onNewTask = { showNewTaskDialog = true },
-                    onClose = onClose,
-                    onCpuExpandedChanged = listener::onTaskManagerCpuExpandedChanged,
-                )
-
-                TaskManagerProcessHeader()
-
-                Box(modifier = Modifier.weight(1f, fill = true).fillMaxWidth()) {
-                    if (taskManagerState.processes.isEmpty()) {
-                        Text(
-                            text = stringResource(R.string.common_ui_no_items_to_display),
-                            color = DrawerTextSecondary,
-                            fontSize = (13f * paneScale).sp,
-                            modifier = Modifier.fillMaxWidth().padding(top = (24f * paneScale).dp),
-                            textAlign = TextAlign.Center,
-                        )
-                    } else {
-                        Column(
-                            modifier =
-                                Modifier
-                                    .fillMaxSize()
-                                    .verticalScroll(rememberScrollState()),
-                            verticalArrangement = Arrangement.spacedBy((12f * paneScale).dp),
-                        ) {
-                            taskManagerState.processes.forEach { process ->
-                                key(process.pid) {
-                                    val selectedAffinityMask =
-                                        pendingAffinities[process.pid]?.affinityMask ?: process.affinityMask
-                                    TaskManagerProcessCard(
-                                        process = process,
-                                        expanded = expandedAffinityPid == process.pid,
-                                        affinityMask = selectedAffinityMask,
-                                        coreCount = affinityCoreCount,
-                                        onToggleAffinity = {
-                                            expandedAffinityPid =
-                                                if (expandedAffinityPid == process.pid) null else process.pid
-                                        },
-                                        onAffinityMaskChanged = { affinityMask ->
-                                            pendingAffinities[process.pid] =
-                                                PendingTaskAffinity(affinityMask, System.currentTimeMillis())
-                                            listener.onTaskManagerSetAffinity(process.pid, affinityMask)
-                                        },
-                                        onEndProcess = { processPendingEnd = process },
-                                        onBringToFront = { listener.onTaskManagerBringToFront(process.name) },
-                                    )
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-    if (showNewTaskDialog) {
-        TaskManagerNewTaskDialog(
-            onDismiss = { showNewTaskDialog = false },
-            onConfirm = { command ->
-                showNewTaskDialog = false
-                listener.onTaskManagerNewTask(command)
-            },
-        )
-    }
-
-    processPendingEnd?.let { process ->
-        TaskManagerEndProcessDialog(
-            process = process,
-            onDismiss = { processPendingEnd = null },
-            onConfirm = {
-                processPendingEnd = null
-                listener.onTaskManagerEndProcess(process.name)
-            },
-        )
-    }
-}
-
-@Composable
-private fun LogsPaneContent(
-    logsState: LogsPaneState,
-    listener: XServerDrawerActionListener,
-    onClose: () -> Unit,
-) {
-    DisposableEffect(Unit) {
-        listener.onLogsPaneVisibilityChanged(true)
-        onDispose { listener.onLogsPaneVisibilityChanged(false) }
-    }
-    BoxWithConstraints(modifier = Modifier.fillMaxSize()) {
-        val paneScale = computePaneScale(maxHeight)
-        CompositionLocalProvider(LocalPaneScale provides paneScale) {
-            Column(
-                modifier =
-                    Modifier
-                        .fillMaxSize()
-                        .padding(horizontal = (12f * paneScale).dp, vertical = (10f * paneScale).dp),
-                verticalArrangement = Arrangement.spacedBy((10f * paneScale).dp),
-            ) {
-                LogsPaneHeader(
-                    paused = logsState.paused,
-                    lineCount = logsState.lines.size,
-                    onClear = { listener.onLogsClear() },
-                    onTogglePause = { listener.onLogsPauseChanged(!logsState.paused) },
-                    onShare = { listener.onLogsShare() },
-                    onClose = onClose,
-                )
-
-                LogsPaneList(
-                    lines = logsState.lines,
-                    paused = logsState.paused,
-                    modifier = Modifier.weight(1f, fill = true).fillMaxWidth(),
-                )
-            }
-        }
-    }
-}
-
-@Composable
-private fun LogsPaneHeader(
-    paused: Boolean,
-    lineCount: Int,
-    onClear: () -> Unit,
-    onTogglePause: () -> Unit,
-    onShare: () -> Unit,
-    onClose: () -> Unit,
-) {
-    val paneScale = LocalPaneScale.current
-    Row(
-        modifier = Modifier.fillMaxWidth(),
-        verticalAlignment = Alignment.CenterVertically,
-        horizontalArrangement = Arrangement.spacedBy((8f * paneScale).dp),
-    ) {
-        Column(modifier = Modifier.weight(1f)) {
-            Text(
-                text = stringResource(R.string.session_drawer_logs),
-                color = DrawerTextPrimary,
-                fontSize = (16f * paneScale).sp,
-                fontWeight = FontWeight.SemiBold,
-            )
-            Text(
-                text =
-                    if (paused) {
-                        stringResource(R.string.session_drawer_logs_paused_indicator) +
-                            " · " +
-                            stringResource(R.string.session_drawer_logs_line_count, lineCount)
-                    } else {
-                        stringResource(R.string.session_drawer_logs_line_count, lineCount)
-                    },
-                color = if (paused) DrawerAccent else DrawerTextSecondary,
-                fontSize = (11f * paneScale).sp,
-                fontWeight = FontWeight.Medium,
-            )
-        }
-
-        LogsPaneActionTile(
-            icon = if (paused) Icons.Outlined.PlayArrow else Icons.Outlined.Pause,
-            contentDescription =
-                if (paused) {
-                    stringResource(R.string.session_drawer_logs_resume)
-                } else {
-                    stringResource(R.string.session_drawer_logs_pause)
-                },
-            onClick = onTogglePause,
-        )
-        LogsPaneActionTile(
-            icon = Icons.Outlined.DeleteSweep,
-            contentDescription = stringResource(R.string.session_drawer_logs_clear),
-            onClick = onClear,
-        )
-        LogsPaneActionTile(
-            icon = Icons.Outlined.Share,
-            contentDescription = stringResource(R.string.session_drawer_logs_share),
-            onClick = onShare,
-        )
-
-        Spacer(Modifier.width((16f * paneScale).dp))
-
-        TaskManagerCloseButton(onClick = onClose)
-    }
-}
-
-@Composable
-private fun LogsPaneActionTile(
-    icon: ImageVector,
-    contentDescription: String,
-    onClick: () -> Unit,
-) {
-    val paneScale = LocalPaneScale.current
-    val interactionSource = remember { MutableInteractionSource() }
-    val pressed = interactionSource.collectIsPressedAsState().value
-    val tint by animateColorAsState(
-        targetValue = if (pressed) DrawerAccent else DrawerTextPrimary,
-        animationSpec = tween(120),
-        label = "logsActionTileTint",
-    )
-    Box(
-        modifier =
-            Modifier
-                .size((38f * paneScale).dp)
-                .clip(CircleShape)
-                .paneNavItem(cornerRadius = (19f * paneScale).dp, onActivate = onClick)
-                .clickable(
-                    interactionSource = interactionSource,
-                    indication = null,
-                    onClick = onClick,
-                ),
-        contentAlignment = Alignment.Center,
-    ) {
-        Icon(
-            imageVector = icon,
-            contentDescription = contentDescription,
-            tint = tint,
-            modifier = Modifier.size((24f * paneScale).dp),
-        )
-    }
-}
-
-@Composable
-private fun LogsPaneList(
-    lines: List<String>,
-    paused: Boolean,
-    modifier: Modifier = Modifier,
-) {
-    val paneScale = LocalPaneScale.current
-    val shape = RoundedCornerShape((10f * paneScale).dp)
-    val listState = rememberLazyListState()
-
-    LaunchedEffect(lines.size, paused) {
-        if (!paused && lines.isNotEmpty()) {
-            listState.scrollToItem((lines.size - 1).coerceAtLeast(0))
-        }
-    }
-
-    Box(
-        modifier =
-            modifier
-                .clip(shape)
-                .background(PaneInnerResting)
-                .border(1.dp, RestingCardBorder, shape),
-    ) {
-        if (lines.isEmpty()) {
-            Text(
-                text = stringResource(R.string.common_ui_no_items_to_display),
-                color = DrawerTextSecondary,
-                fontSize = (12f * paneScale).sp,
-                modifier =
-                    Modifier
-                        .fillMaxWidth()
-                        .padding(top = (24f * paneScale).dp),
-                textAlign = TextAlign.Center,
-            )
-        } else {
-            LazyColumn(
-                state = listState,
-                modifier = Modifier.fillMaxSize(),
-                contentPadding =
-                    PaddingValues(
-                        horizontal = (10f * paneScale).dp,
-                        vertical = (8f * paneScale).dp,
-                    ),
-                verticalArrangement = Arrangement.spacedBy((1f * paneScale).dp),
-            ) {
-                items(lines) { line ->
-                    Text(
-                        text = line,
-                        color = DrawerTextPrimary,
-                        fontSize = (11f * paneScale).sp,
-                        fontFamily = FontFamily.Monospace,
-                        lineHeight = (14f * paneScale).sp,
-                        letterSpacing = 0.sp,
-                        modifier = Modifier.fillMaxWidth(),
-                    )
-                }
-            }
-        }
-    }
-}
-
-@Composable
-private fun TaskManagerHeader(
-    cpuPercent: Int,
-    cpuCoreCount: Int,
-    cpuCorePercents: List<Int>,
-    memoryPercent: Int,
-    memoryDetail: String,
-    onNewTask: () -> Unit,
-    onClose: () -> Unit,
-    onCpuExpandedChanged: (Boolean) -> Unit,
-) {
-    val paneScale = LocalPaneScale.current
-    var cpuExpanded by remember { mutableStateOf(false) }
-    DisposableEffect(cpuExpanded) {
-        onCpuExpandedChanged(cpuExpanded)
-        onDispose { if (cpuExpanded) onCpuExpandedChanged(false) }
-    }
-
-    Row(
-        modifier = Modifier.fillMaxWidth(),
-        verticalAlignment = Alignment.CenterVertically,
-    ) {
-        Text(
-            text = stringResource(R.string.session_task_title),
-            color = DrawerTextPrimary,
-            fontSize = (16f * paneScale).sp,
-            fontWeight = FontWeight.SemiBold,
-            modifier = Modifier.weight(1f),
-        )
-
-        TaskManagerCloseButton(onClick = onClose)
-    }
-
-    Row(
-        modifier = Modifier.fillMaxWidth().height(IntrinsicSize.Max),
-        horizontalArrangement = Arrangement.spacedBy((8f * paneScale).dp),
-        verticalAlignment = Alignment.CenterVertically,
-    ) {
-        TaskManagerStatTile(
-            title = stringResource(R.string.session_task_cpu_usage_format, cpuPercent),
-            detail =
-                if (cpuCoreCount > 0) {
-                    pluralStringResource(R.plurals.session_task_core_count, cpuCoreCount, cpuCoreCount)
-                } else {
-                    ""
-                },
-            modifier = Modifier.weight(1f).fillMaxHeight(),
-            selected = cpuExpanded,
-            onClick = { cpuExpanded = !cpuExpanded },
-        )
-        TaskManagerStatTile(
-            title = stringResource(R.string.session_task_memory) + " ($memoryPercent%)",
-            detail = memoryDetail,
-            modifier = Modifier.weight(1f).fillMaxHeight(),
-        )
-    }
-
-    AnimatedVisibility(
-        visible = cpuExpanded && cpuCorePercents.isNotEmpty(),
-        enter =
-            fadeIn(animationSpec = tween(durationMillis = 180, easing = FastOutSlowInEasing)) +
-                expandVertically(
-                    animationSpec = tween(durationMillis = 220, easing = FastOutSlowInEasing),
-                    expandFrom = Alignment.Top,
-                ),
-        exit =
-            fadeOut(animationSpec = tween(durationMillis = 140, easing = FastOutSlowInEasing)) +
-                shrinkVertically(
-                    animationSpec = tween(durationMillis = 180, easing = FastOutSlowInEasing),
-                    shrinkTowards = Alignment.Top,
-                ),
-    ) {
-        TaskManagerCpuCoreGrid(cpuCorePercents = cpuCorePercents)
-    }
-
-    TaskManagerNewTaskButton(onClick = onNewTask)
-}
-
-@Composable
-private fun TaskManagerCloseButton(onClick: () -> Unit) {
-    val paneScale = LocalPaneScale.current
-    val interactionSource = remember { MutableInteractionSource() }
-    val pressed = interactionSource.collectIsPressedAsState().value
-    val bgColor by animateColorAsState(
-        targetValue = if (pressed) PaneInnerPressed else PaneInnerResting,
-        animationSpec = tween(120),
-        label = "taskManagerCloseBg",
-    )
-    val size = (38f * paneScale).dp
-    val shape = RoundedCornerShape((10f * paneScale).dp)
-    Box(
-        modifier =
-            Modifier
-                .size(size)
-                .clip(shape)
-                .background(bgColor)
-                .border(1.dp, RestingCardBorder, shape)
-                .paneNavItem(cornerRadius = (10f * paneScale).dp, onActivate = onClick)
-                .clickable(
-                    interactionSource = interactionSource,
-                    indication = null,
-                    onClick = onClick,
-                ),
-        contentAlignment = Alignment.Center,
-    ) {
-        Icon(
-            imageVector = Icons.Outlined.Close,
-            contentDescription = stringResource(R.string.common_ui_close),
-            tint = DrawerTextPrimary,
-            modifier = Modifier.size((22f * paneScale).dp),
-        )
-    }
-}
-
-@Composable
-private fun TaskManagerStatTile(
-    title: String,
-    detail: String,
-    modifier: Modifier = Modifier,
-    selected: Boolean = false,
-    onClick: (() -> Unit)? = null,
-) {
-    val paneScale = LocalPaneScale.current
-    val shape = RoundedCornerShape((10f * paneScale).dp)
-    val interactionSource = remember { MutableInteractionSource() }
-    val pressed = interactionSource.collectIsPressedAsState().value
-    val bgColor by animateColorAsState(
-        targetValue =
-            when {
-                pressed -> PaneInnerPressed
-                else -> PaneInnerResting
-            },
-        animationSpec = tween(120),
-        label = "taskManagerStatTileBg",
-    )
-    val borderColor by animateColorAsState(
-        targetValue = if (selected) DrawerAccent else RestingCardBorder,
-        animationSpec = tween(120),
-        label = "taskManagerStatTileBorder",
-    )
-    val clickModifier =
-        if (onClick != null) {
-            Modifier.clickable(
-                interactionSource = interactionSource,
-                indication = null,
-                onClick = onClick,
-            )
-        } else {
-            Modifier
-        }
-
-    Column(
-        modifier =
-            modifier
-                .clip(shape)
-                .background(bgColor)
-                .border(1.dp, borderColor, shape)
-                .then(clickModifier)
-                .padding(horizontal = (8f * paneScale).dp, vertical = (6f * paneScale).dp),
-        verticalArrangement = Arrangement.spacedBy((1f * paneScale).dp),
-    ) {
-        Text(
-            text = title,
-            color = DrawerAccent,
-            fontSize = (11f * paneScale).sp,
-            fontWeight = FontWeight.SemiBold,
-            maxLines = 1,
-            overflow = TextOverflow.Ellipsis,
-            textAlign = TextAlign.Center,
-            modifier = Modifier.fillMaxWidth(),
-        )
-        Text(
-            text = detail,
-            color = DrawerTextSecondary,
-            fontSize = (10f * paneScale).sp,
-            maxLines = 1,
-            overflow = TextOverflow.Ellipsis,
-            textAlign = TextAlign.Center,
-            modifier = Modifier.fillMaxWidth(),
-        )
-    }
-}
-
-@OptIn(ExperimentalLayoutApi::class)
-@Composable
-private fun TaskManagerCpuCoreGrid(cpuCorePercents: List<Int>) {
-    val paneScale = LocalPaneScale.current
-    val shape = RoundedCornerShape((10f * paneScale).dp)
-    Column(
-        modifier =
-            Modifier
-                .fillMaxWidth()
-                .clip(shape)
-                .background(PaneInnerResting)
-                .border(1.dp, RestingCardBorder, shape)
-                .padding(horizontal = (8f * paneScale).dp, vertical = (6f * paneScale).dp),
-        verticalArrangement = Arrangement.spacedBy((4f * paneScale).dp),
-    ) {
-        Text(
-            text = stringResource(R.string.session_task_per_core_usage),
-            color = DrawerTextPrimary,
-            fontSize = (11f * paneScale).sp,
-            fontWeight = FontWeight.SemiBold,
-        )
-        FlowRow(
-            modifier = Modifier.fillMaxWidth(),
-            horizontalArrangement = Arrangement.spacedBy((4f * paneScale).dp),
-            verticalArrangement = Arrangement.spacedBy((4f * paneScale).dp),
-        ) {
-            cpuCorePercents.forEachIndexed { index, percent ->
-                TaskManagerCpuCoreChip(coreIndex = index, percent = percent)
-            }
-        }
-    }
-}
-
-@Composable
-private fun TaskManagerCpuCoreChip(coreIndex: Int, percent: Int) {
-    val paneScale = LocalPaneScale.current
-    val shape = RoundedCornerShape((6f * paneScale).dp)
-    Row(
-        modifier =
-            Modifier
-                .clip(shape)
-                .background(PaneSurfaceColor)
-                .border(1.dp, RestingCardBorder, shape)
-                .padding(horizontal = (6f * paneScale).dp, vertical = (3f * paneScale).dp),
-        verticalAlignment = Alignment.CenterVertically,
-        horizontalArrangement = Arrangement.spacedBy((4f * paneScale).dp),
-    ) {
-        Text(
-            text = stringResource(R.string.session_task_core_label, coreIndex),
-            color = DrawerTextSecondary,
-            fontSize = (10f * paneScale).sp,
-            fontWeight = FontWeight.Medium,
-        )
-        Text(
-            text = "$percent%",
-            color = DrawerAccent,
-            fontSize = (10f * paneScale).sp,
-            fontWeight = FontWeight.SemiBold,
-        )
-    }
-}
-
-@Composable
-private fun TaskManagerNewTaskButton(onClick: () -> Unit) {
-    val paneScale = LocalPaneScale.current
-    val interactionSource = remember { MutableInteractionSource() }
-    val pressed = interactionSource.collectIsPressedAsState().value
-    val bgColor by animateColorAsState(
-        targetValue = if (pressed) PaneInnerPressed else PaneInnerResting,
-        animationSpec = tween(120),
-        label = "taskManagerNewTaskBg",
-    )
-    val tint = if (pressed) DrawerAccent.copy(alpha = 0.76f) else DrawerAccent
-    val shape = RoundedCornerShape((12f * paneScale).dp)
-    Row(
-        modifier =
-            Modifier
-                .fillMaxWidth()
-                .clip(shape)
-                .background(bgColor)
-                .border(1.dp, RestingCardBorder, shape)
-                .paneNavItem(cornerRadius = (12f * paneScale).dp, onActivate = onClick)
-                .clickable(
-                    interactionSource = interactionSource,
-                    indication = null,
-                    onClick = onClick,
-                )
-                .padding(horizontal = (12f * paneScale).dp, vertical = (10f * paneScale).dp),
-        horizontalArrangement = Arrangement.Center,
-        verticalAlignment = Alignment.CenterVertically,
-    ) {
-        Icon(
-            imageVector = Icons.Outlined.Add,
-            contentDescription = null,
-            tint = tint,
-            modifier = Modifier.size((18f * paneScale).dp),
-        )
-        Spacer(Modifier.width((6f * paneScale).dp))
-        Text(
-            text = stringResource(R.string.session_task_new_task),
-            color = tint,
-            fontSize = (14f * paneScale).sp,
-            fontWeight = FontWeight.Medium,
-        )
-    }
-}
-
-@OptIn(ExperimentalLayoutApi::class)
-@Composable
-private fun TaskManagerAffinityOptions(
-    affinityMask: Int,
-    coreCount: Int,
-    onAffinityMaskChanged: (Int) -> Unit,
-) {
-    val paneScale = LocalPaneScale.current
-    val effectiveCoreCount = coreCount.coerceAtLeast(1).coerceAtMost(32)
-    val selectedMask = sanitizeTaskAffinityMask(affinityMask, effectiveCoreCount)
-    val fullMask = taskAffinityFullMask(effectiveCoreCount)
-
-    Column(
-        modifier =
-            Modifier
-                .fillMaxWidth()
-                .padding(
-                    start = (8f * paneScale).dp,
-                    end = (8f * paneScale).dp,
-                    bottom = (8f * paneScale).dp,
-                ),
-        verticalArrangement = Arrangement.spacedBy((7f * paneScale).dp),
-    ) {
-        Row(
-            modifier = Modifier.fillMaxWidth(),
-            verticalAlignment = Alignment.CenterVertically,
-            horizontalArrangement = Arrangement.spacedBy((6f * paneScale).dp),
-        ) {
-            Icon(
-                imageVector = Icons.Outlined.Tune,
-                contentDescription = null,
-                tint = DrawerAccent,
-                modifier = Modifier.size((15f * paneScale).dp),
-            )
-            Text(
-                text = stringResource(R.string.session_task_affinity_title),
-                color = DrawerTextPrimary,
-                fontSize = (12f * paneScale).sp,
-                fontWeight = FontWeight.SemiBold,
-                modifier = Modifier.weight(1f),
-            )
-        }
-
-        FlowRow(
-            modifier = Modifier.fillMaxWidth(),
-            horizontalArrangement = Arrangement.spacedBy((5f * paneScale).dp),
-            verticalArrangement = Arrangement.spacedBy((5f * paneScale).dp),
-        ) {
-            TaskManagerAffinityChip(
-                label = stringResource(R.string.session_task_affinity_all_cores),
-                selected = selectedMask == fullMask,
-                onClick = { onAffinityMaskChanged(fullMask) },
-            )
-            for (coreIndex in 0 until effectiveCoreCount) {
-                val bit = 1 shl coreIndex
-                TaskManagerAffinityChip(
-                    label = stringResource(R.string.session_task_core_label, coreIndex),
-                    selected = (selectedMask and bit) != 0,
-                    onClick = {
-                        val nextMask =
-                            if ((selectedMask and bit) != 0) {
-                                selectedMask and bit.inv()
-                            } else {
-                                selectedMask or bit
-                            }
-                        if ((nextMask and fullMask) != 0) {
-                            onAffinityMaskChanged(nextMask and fullMask)
-                        }
-                    },
-                )
-            }
-        }
-    }
-}
-
-@Composable
-private fun TaskManagerAffinityChip(
-    label: String,
-    selected: Boolean,
-    onClick: () -> Unit,
-) {
-    val bgColor =
-        if (selected) {
-            DrawerAccent.copy(alpha = 0.16f)
-        } else {
-            PaneInnerResting
-        }
-    val borderColor = if (selected) DrawerAccent.copy(alpha = 0.56f) else RestingCardBorder
-    val textColor = if (selected) DrawerAccent else DrawerTextPrimary
-    Row(
-        modifier =
-            Modifier
-                .clip(RoundedCornerShape(8.dp))
-                .background(bgColor)
-                .border(1.dp, borderColor, RoundedCornerShape(8.dp))
-                .paneNavItem(cornerRadius = 8.dp, onActivate = onClick)
-                .clickable(
-                    interactionSource = remember { MutableInteractionSource() },
-                    indication = null,
-                    onClick = onClick,
-                )
-                .padding(horizontal = 8.dp, vertical = 5.dp),
-        verticalAlignment = Alignment.CenterVertically,
-        horizontalArrangement = Arrangement.spacedBy(4.dp),
-    ) {
-        Icon(
-            imageVector = Icons.Outlined.Check,
-            contentDescription = null,
-            tint = DrawerAccent.copy(alpha = if (selected) 1f else 0f),
-            modifier = Modifier.size(13.dp),
-        )
-        Text(
-            text = label,
-            color = textColor,
-            fontSize = 11.sp,
-            fontWeight = if (selected) FontWeight.SemiBold else FontWeight.Medium,
-        )
-    }
-}
-
-private fun taskAffinityFullMask(coreCount: Int): Int {
-    var mask = 0
-    for (index in 0 until coreCount.coerceAtLeast(1).coerceAtMost(32)) {
-        mask = mask or (1 shl index)
-    }
-    return mask
-}
-
-private fun sanitizeTaskAffinityMask(affinityMask: Int, coreCount: Int): Int {
-    val fullMask = taskAffinityFullMask(coreCount)
-    val sanitizedMask = affinityMask and fullMask
-    return if (sanitizedMask != 0) sanitizedMask else fullMask
-}
-
-private class TaskManagerPopupPositionProvider(private val gapPx: Int) : PopupPositionProvider {
+internal class TaskManagerPopupPositionProvider(private val gapPx: Int) : PopupPositionProvider {
     override fun calculatePosition(
         anchorBounds: IntRect,
         windowSize: IntSize,
@@ -4731,627 +2378,10 @@ private class TaskManagerPopupPositionProvider(private val gapPx: Int) : PopupPo
     }
 }
 
-@Composable
-private fun TaskManagerActionPopup(
-    expanded: Boolean,
-    onDismiss: () -> Unit,
-    content: @Composable ColumnScope.() -> Unit,
-) {
-    if (!expanded) return
-    val paneScale = LocalPaneScale.current
-    val density = LocalDensity.current
-    val gapPx = with(density) { (4f * paneScale).dp.roundToPx() }
-    val shape = RoundedCornerShape((12f * paneScale).dp)
-    Popup(
-        popupPositionProvider = remember(gapPx) { TaskManagerPopupPositionProvider(gapPx) },
-        onDismissRequest = onDismiss,
-        properties = PopupProperties(focusable = true),
-    ) {
-        Column(
-            modifier =
-                Modifier
-                    .width(IntrinsicSize.Max)
-                    .widthIn(min = (150f * paneScale).dp, max = (240f * paneScale).dp)
-                    .clip(shape)
-                    .background(PaneSurfaceColor)
-                    .border(1.dp, RestingCardBorder, shape)
-                    .padding((5f * paneScale).dp),
-            verticalArrangement = Arrangement.spacedBy((4f * paneScale).dp),
-            content = content,
-        )
-    }
-}
+internal val NEW_TASK_PRESETS = listOf("Wfm.exe", "Winecfg.exe", "Regedit.exe", "Taskmgr.exe", "Services.exe")
 
 @Composable
-private fun TaskManagerActionPopupItem(
-    label: String,
-    onClick: () -> Unit,
-    icon: ImageVector? = null,
-) {
-    val paneScale = LocalPaneScale.current
-    val interactionSource = remember { MutableInteractionSource() }
-    val pressed = interactionSource.collectIsPressedAsState().value
-    val bgColor by animateColorAsState(
-        targetValue = if (pressed) DrawerAccent.copy(alpha = 0.16f) else PaneInnerResting,
-        animationSpec = tween(120),
-        label = "taskManagerPopupItem",
-    )
-    val shape = RoundedCornerShape((8f * paneScale).dp)
-    Row(
-        modifier =
-            Modifier
-                .fillMaxWidth()
-                .clip(shape)
-                .background(bgColor)
-                .border(1.dp, RestingCardBorder, shape)
-                .clickable(interactionSource = interactionSource, indication = null, onClick = onClick)
-                .padding(horizontal = (12f * paneScale).dp, vertical = (10f * paneScale).dp),
-        verticalAlignment = Alignment.CenterVertically,
-        horizontalArrangement = Arrangement.spacedBy((8f * paneScale).dp),
-    ) {
-        if (icon != null) {
-            Icon(
-                imageVector = icon,
-                contentDescription = null,
-                tint = DrawerAccent,
-                modifier = Modifier.size((16f * paneScale).dp),
-            )
-        }
-        Text(
-            text = label,
-            color = DrawerTextPrimary,
-            fontSize = (13f * paneScale).sp,
-            fontWeight = FontWeight.Medium,
-            maxLines = 1,
-            overflow = TextOverflow.Ellipsis,
-        )
-    }
-}
-
-@Composable
-private fun TaskManagerEndProcessDialog(
-    process: TaskManagerProcess,
-    onDismiss: () -> Unit,
-    onConfirm: () -> Unit,
-) {
-    val displayName = if (process.isWow64) "${process.name} *32" else process.name
-    val shape = RoundedCornerShape(12.dp)
-
-    Dialog(
-        onDismissRequest = onDismiss,
-        properties =
-            DialogProperties(
-                usePlatformDefaultWidth = false,
-                decorFitsSystemWindows = false,
-            ),
-    ) {
-        Box(
-            modifier =
-                Modifier
-                    .fillMaxSize()
-                    .safeDrawingPadding()
-                    .padding(horizontal = 14.dp, vertical = 10.dp),
-            contentAlignment = Alignment.Center,
-        ) {
-            Column(
-                modifier =
-                    Modifier
-                        .widthIn(max = 292.dp)
-                        .fillMaxWidth()
-                        .clip(shape)
-                        .background(PaneSurfaceColor)
-                        .border(1.dp, GlassExitTint.copy(alpha = 0.32f), shape)
-                        .padding(horizontal = 12.dp, vertical = 10.dp),
-                verticalArrangement = Arrangement.spacedBy(8.dp),
-            ) {
-                Row(
-                    modifier = Modifier.fillMaxWidth(),
-                    verticalAlignment = Alignment.CenterVertically,
-                    horizontalArrangement = Arrangement.spacedBy(8.dp),
-                ) {
-                    Icon(
-                        imageVector = Icons.Outlined.Close,
-                        contentDescription = null,
-                        tint = GlassExitTint,
-                        modifier = Modifier.size(17.dp),
-                    )
-                    Text(
-                        text = stringResource(R.string.session_task_end_process),
-                        color = DrawerTextPrimary,
-                        fontSize = 13.sp,
-                        fontWeight = FontWeight.SemiBold,
-                        maxLines = 1,
-                        overflow = TextOverflow.Ellipsis,
-                        modifier = Modifier.weight(1f),
-                    )
-                }
-
-                Text(
-                    text = displayName,
-                    color = DrawerTextPrimary,
-                    fontSize = 12.sp,
-                    fontWeight = FontWeight.Medium,
-                    maxLines = 1,
-                    overflow = TextOverflow.Ellipsis,
-                    textAlign = TextAlign.Center,
-                    modifier = Modifier.fillMaxWidth(),
-                )
-                Text(
-                    text = stringResource(R.string.session_task_confirm_end_process),
-                    color = DrawerTextPrimary,
-                    fontSize = 11.sp,
-                    lineHeight = 14.sp,
-                    textAlign = TextAlign.Center,
-                    modifier = Modifier.fillMaxWidth().padding(bottom = 4.dp),
-                )
-
-                Row(
-                    modifier = Modifier.fillMaxWidth(),
-                    horizontalArrangement = Arrangement.spacedBy(8.dp, Alignment.End),
-                    verticalAlignment = Alignment.CenterVertically,
-                ) {
-                    TaskManagerDialogButton(
-                        label = stringResource(R.string.common_ui_cancel),
-                        textColor = DrawerTextPrimary,
-                        modifier = Modifier.height(34.dp),
-                        verticalPadding = 0.dp,
-                        onClick = onDismiss,
-                    )
-                    TaskManagerDialogButton(
-                        label = stringResource(R.string.session_task_end_process),
-                        textColor = GlassExitTint,
-                        modifier = Modifier.height(34.dp),
-                        verticalPadding = 0.dp,
-                        fontWeight = FontWeight.Medium,
-                        backgroundColor = GlassExitTint.copy(alpha = 0.12f),
-                        borderColor = GlassExitTint.copy(alpha = 0.34f),
-                        onClick = onConfirm,
-                    )
-                }
-            }
-        }
-    }
-}
-
-private val NEW_TASK_PRESETS = listOf("Wfm.exe", "Winecfg.exe", "Regedit.exe", "Taskmgr.exe", "Services.exe")
-
-@Composable
-private fun TaskManagerNewTaskDialog(
-    onDismiss: () -> Unit,
-    onConfirm: (String) -> Unit,
-) {
-    val customLabel = stringResource(R.string.session_task_custom_value)
-    var selectedLabel by remember { mutableStateOf(NEW_TASK_PRESETS.first()) }
-    var customMode by remember { mutableStateOf(false) }
-    var customText by remember { mutableStateOf("") }
-    var dropdownExpanded by remember { mutableStateOf(false) }
-    val focusRequester = remember { FocusRequester() }
-    val keyboardController = LocalSoftwareKeyboardController.current
-    val shape = RoundedCornerShape(14.dp)
-    val fieldShape = RoundedCornerShape(10.dp)
-
-    fun submit() {
-        val command = if (customMode) customText.trim() else selectedLabel.trim().lowercase()
-        if (command.isNotEmpty()) onConfirm(command)
-    }
-
-    LaunchedEffect(customMode) {
-        if (customMode) {
-            focusRequester.requestFocus()
-            keyboardController?.show()
-        }
-    }
-
-    Dialog(
-        onDismissRequest = onDismiss,
-        properties =
-            DialogProperties(
-                usePlatformDefaultWidth = false,
-                decorFitsSystemWindows = false,
-            ),
-    ) {
-        Box(
-            modifier =
-                Modifier
-                    .fillMaxSize()
-                    .safeDrawingPadding()
-                    .imePadding()
-                    .padding(horizontal = 14.dp, vertical = 10.dp),
-            contentAlignment = Alignment.Center,
-        ) {
-            Column(
-                modifier =
-                    Modifier
-                        .widthIn(max = 310.dp)
-                        .fillMaxWidth()
-                        .clip(shape)
-                        .background(PaneSurfaceColor)
-                        .border(1.dp, RestingCardBorder, shape)
-                        .padding(horizontal = 14.dp, vertical = 12.dp),
-                verticalArrangement = Arrangement.spacedBy(10.dp),
-            ) {
-                Row(
-                    modifier = Modifier.fillMaxWidth(),
-                    verticalAlignment = Alignment.CenterVertically,
-                    horizontalArrangement = Arrangement.spacedBy(8.dp),
-                ) {
-                    Icon(
-                        imageVector = Icons.Outlined.Add,
-                        contentDescription = null,
-                        tint = DrawerAccent,
-                        modifier = Modifier.size(18.dp),
-                    )
-                    Text(
-                        text = stringResource(R.string.session_task_new_task),
-                        color = DrawerTextPrimary,
-                        fontSize = 14.sp,
-                        fontWeight = FontWeight.SemiBold,
-                    )
-                }
-
-                Box(modifier = Modifier.fillMaxWidth()) {
-                    if (customMode) {
-                        OutlinedTextField(
-                            value = customText,
-                            onValueChange = { customText = it },
-                            modifier =
-                                Modifier
-                                    .fillMaxWidth()
-                                    .height(48.dp)
-                                    .focusRequester(focusRequester),
-                            singleLine = true,
-                            textStyle =
-                                androidx.compose.material3.MaterialTheme.typography.bodyMedium.copy(
-                                    color = DrawerTextPrimary,
-                                    fontSize = 13.sp,
-                                ),
-                            colors =
-                                OutlinedTextFieldDefaults.colors(
-                                    focusedBorderColor = DrawerAccent,
-                                    unfocusedBorderColor = RestingCardBorder,
-                                    focusedTextColor = DrawerTextPrimary,
-                                    unfocusedTextColor = DrawerTextPrimary,
-                                    focusedContainerColor = PaneInnerResting,
-                                    unfocusedContainerColor = PaneInnerResting,
-                                    cursorColor = DrawerAccent,
-                                ),
-                            trailingIcon = {
-                                Icon(
-                                    imageVector = Icons.Outlined.ArrowDropDown,
-                                    contentDescription = null,
-                                    tint = DrawerTextSecondary,
-                                    modifier =
-                                        Modifier
-                                            .size(22.dp)
-                                            .clickable(
-                                                interactionSource = remember { MutableInteractionSource() },
-                                                indication = null,
-                                            ) { dropdownExpanded = true },
-                                )
-                            },
-                            keyboardOptions = KeyboardOptions(imeAction = ImeAction.Done),
-                            keyboardActions =
-                                KeyboardActions(
-                                    onDone = {
-                                        keyboardController?.hide()
-                                        submit()
-                                    },
-                                ),
-                        )
-                    } else {
-                        Row(
-                            modifier =
-                                Modifier
-                                    .fillMaxWidth()
-                                    .clip(fieldShape)
-                                    .background(PaneInnerResting)
-                                    .border(1.dp, RestingCardBorder, fieldShape)
-                                    .clickable { dropdownExpanded = true }
-                                    .padding(horizontal = 12.dp, vertical = 12.dp),
-                            verticalAlignment = Alignment.CenterVertically,
-                        ) {
-                            Text(
-                                text = selectedLabel,
-                                color = DrawerTextPrimary,
-                                fontSize = 13.sp,
-                                modifier = Modifier.weight(1f),
-                            )
-                            Icon(
-                                imageVector = Icons.Outlined.ArrowDropDown,
-                                contentDescription = null,
-                                tint = DrawerTextSecondary,
-                                modifier = Modifier.size(22.dp),
-                            )
-                        }
-                    }
-                    TaskManagerActionPopup(
-                        expanded = dropdownExpanded,
-                        onDismiss = { dropdownExpanded = false },
-                    ) {
-                        NEW_TASK_PRESETS.forEach { item ->
-                            TaskManagerActionPopupItem(
-                                label = item,
-                                onClick = {
-                                    selectedLabel = item
-                                    customMode = false
-                                    dropdownExpanded = false
-                                },
-                            )
-                        }
-                        TaskManagerActionPopupItem(
-                            label = customLabel,
-                            onClick = {
-                                customMode = true
-                                dropdownExpanded = false
-                            },
-                        )
-                    }
-                }
-
-                Row(
-                    modifier = Modifier.fillMaxWidth(),
-                    horizontalArrangement = Arrangement.spacedBy(8.dp, Alignment.End),
-                    verticalAlignment = Alignment.CenterVertically,
-                ) {
-                    TaskManagerDialogButton(
-                        label = stringResource(R.string.common_ui_cancel),
-                        textColor = DrawerTextPrimary,
-                        modifier = Modifier.height(34.dp),
-                        verticalPadding = 0.dp,
-                        onClick = onDismiss,
-                    )
-                    TaskManagerDialogButton(
-                        label = stringResource(R.string.common_ui_ok),
-                        textColor = DrawerAccent,
-                        modifier = Modifier.height(34.dp),
-                        verticalPadding = 0.dp,
-                        backgroundColor = DrawerAccent.copy(alpha = 0.12f),
-                        borderColor = DrawerAccent.copy(alpha = 0.34f),
-                        onClick = {
-                            keyboardController?.hide()
-                            submit()
-                        },
-                    )
-                }
-            }
-        }
-    }
-}
-
-@Composable
-private fun TaskManagerDialogButton(
-    label: String,
-    textColor: Color,
-    onClick: () -> Unit,
-    modifier: Modifier = Modifier,
-    backgroundColor: Color = PaneInnerResting,
-    borderColor: Color = RestingCardBorder,
-    fontWeight: FontWeight = FontWeight.SemiBold,
-    verticalPadding: Dp = 8.dp,
-) {
-    Box(
-        modifier =
-            modifier
-                .widthIn(min = 72.dp)
-                .clip(RoundedCornerShape(9.dp))
-                .background(backgroundColor)
-                .border(1.dp, borderColor, RoundedCornerShape(9.dp))
-                .clickable(
-                    interactionSource = remember { MutableInteractionSource() },
-                    indication = null,
-                    onClick = onClick,
-                )
-                .padding(horizontal = 14.dp, vertical = verticalPadding),
-        contentAlignment = Alignment.Center,
-    ) {
-        Text(
-            text = label,
-            color = textColor,
-            fontSize = 12.sp,
-            fontWeight = fontWeight,
-            maxLines = 1,
-            overflow = TextOverflow.Ellipsis,
-        )
-    }
-}
-
-@Composable
-private fun TaskManagerProcessHeader() {
-    val paneScale = LocalPaneScale.current
-    Row(
-        modifier = Modifier.fillMaxWidth().padding(horizontal = (4f * paneScale).dp),
-        verticalAlignment = Alignment.CenterVertically,
-    ) {
-        Text(
-            text = stringResource(R.string.session_task_process_name),
-            color = DrawerTextSecondary,
-            fontSize = (11f * paneScale).sp,
-            fontWeight = FontWeight.Medium,
-            modifier = Modifier.weight(1f),
-        )
-        Text(
-            text = stringResource(R.string.session_task_pid),
-            color = DrawerTextSecondary,
-            fontSize = (11f * paneScale).sp,
-            fontWeight = FontWeight.Medium,
-            textAlign = TextAlign.End,
-            modifier = Modifier.width((54f * paneScale).dp),
-        )
-        Text(
-            text = stringResource(R.string.session_task_memory),
-            color = DrawerTextSecondary,
-            fontSize = (11f * paneScale).sp,
-            fontWeight = FontWeight.Medium,
-            textAlign = TextAlign.End,
-            modifier = Modifier.width((78f * paneScale).dp),
-        )
-        Spacer(modifier = Modifier.width((46f * paneScale).dp))
-    }
-}
-
-@OptIn(ExperimentalFoundationApi::class)
-@Composable
-private fun TaskManagerProcessCard(
-    process: TaskManagerProcess,
-    expanded: Boolean,
-    affinityMask: Int,
-    coreCount: Int,
-    onToggleAffinity: () -> Unit,
-    onAffinityMaskChanged: (Int) -> Unit,
-    onEndProcess: () -> Unit,
-    onBringToFront: () -> Unit,
-) {
-    val paneScale = LocalPaneScale.current
-    val shape = RoundedCornerShape((8f * paneScale).dp)
-    val displayName = if (process.isWow64) "${process.name} *32" else process.name
-    val interactionSource = remember { MutableInteractionSource() }
-    val pressed = interactionSource.collectIsPressedAsState().value
-    var menuExpanded by remember { mutableStateOf(false) }
-    val bgColor by animateColorAsState(
-        targetValue = if (pressed) PaneInnerPressed else PaneInnerResting,
-        animationSpec = tween(120),
-        label = "taskManagerProcessRowBg",
-    )
-    val borderColor by animateColorAsState(
-        targetValue = if (expanded) DrawerAccent.copy(alpha = 0.62f) else RestingCardBorder,
-        animationSpec = tween(160),
-        label = "taskManagerProcessCardBorder",
-    )
-
-    Column(
-        modifier =
-            Modifier
-                .fillMaxWidth()
-                .clip(shape)
-                .background(bgColor)
-                .border(1.dp, borderColor, shape),
-    ) {
-        Box {
-            Row(
-                modifier =
-                    Modifier
-                        .fillMaxWidth()
-                        .paneNavItem(
-                            cornerRadius = (8f * paneScale).dp,
-                            onActivate = onToggleAffinity,
-                            onSecondary = onBringToFront,
-                        )
-                        .combinedClickable(
-                            interactionSource = interactionSource,
-                            indication = null,
-                            onClick = onToggleAffinity,
-                            onLongClick = { menuExpanded = true },
-                        )
-                        .padding(horizontal = (8f * paneScale).dp, vertical = (6f * paneScale).dp),
-                verticalAlignment = Alignment.CenterVertically,
-            ) {
-                Text(
-                    text = displayName,
-                    color = DrawerTextPrimary,
-                    fontSize = (12f * paneScale).sp,
-                    fontWeight = FontWeight.Medium,
-                    maxLines = 1,
-                    overflow = TextOverflow.Ellipsis,
-                    modifier = Modifier.weight(1f),
-                )
-                Text(
-                    text = process.pid.toString(),
-                    color = DrawerTextSecondary,
-                    fontSize = (12f * paneScale).sp,
-                    textAlign = TextAlign.End,
-                    modifier = Modifier.width((54f * paneScale).dp),
-                )
-                Text(
-                    text = process.memoryFormatted,
-                    color = DrawerTextSecondary,
-                    fontSize = (12f * paneScale).sp,
-                    textAlign = TextAlign.End,
-                    modifier = Modifier.width((78f * paneScale).dp),
-                )
-                Spacer(modifier = Modifier.width((10f * paneScale).dp))
-                Box(
-                    Modifier.paneNavItem(cornerRadius = (8f * paneScale).dp, onActivate = onEndProcess),
-                ) {
-                    TaskManagerEndButton(onClick = onEndProcess)
-                }
-            }
-
-            TaskManagerActionPopup(
-                expanded = menuExpanded,
-                onDismiss = { menuExpanded = false },
-            ) {
-                TaskManagerActionPopupItem(
-                    label = stringResource(R.string.session_task_bring_to_front),
-                    icon = Icons.Outlined.Monitor,
-                    onClick = {
-                        menuExpanded = false
-                        onBringToFront()
-                    },
-                )
-            }
-        }
-
-        AnimatedVisibility(
-            visible = expanded,
-            enter =
-                fadeIn(animationSpec = tween(durationMillis = 150, easing = FastOutSlowInEasing)) +
-                    expandVertically(
-                        animationSpec = tween(durationMillis = 220, easing = FastOutSlowInEasing),
-                        expandFrom = Alignment.Top,
-                    ),
-            exit =
-                fadeOut(animationSpec = tween(durationMillis = 120, easing = FastOutSlowInEasing)) +
-                    shrinkVertically(
-                        animationSpec = tween(durationMillis = 180, easing = FastOutSlowInEasing),
-                        shrinkTowards = Alignment.Top,
-                    ),
-        ) {
-            TaskManagerAffinityOptions(
-                affinityMask = affinityMask,
-                coreCount = coreCount,
-                onAffinityMaskChanged = onAffinityMaskChanged,
-            )
-        }
-    }
-}
-
-@Composable
-private fun TaskManagerEndButton(onClick: () -> Unit) {
-    val paneScale = LocalPaneScale.current
-    val interactionSource = remember { MutableInteractionSource() }
-    val pressed = interactionSource.collectIsPressedAsState().value
-    val bgColor by animateColorAsState(
-        targetValue = if (pressed) TileExitPressed else TileExitResting,
-        animationSpec = tween(120),
-        label = "taskManagerEndBtn",
-    )
-    val size = (32f * paneScale).dp
-    val shape = RoundedCornerShape((8f * paneScale).dp)
-    Box(
-        modifier =
-            Modifier
-                .size(size)
-                .clip(shape)
-                .background(bgColor)
-                .border(1.dp, GlassExitTint.copy(alpha = 0.34f), shape)
-                .clickable(
-                    interactionSource = interactionSource,
-                    indication = null,
-                    onClick = onClick,
-                ),
-        contentAlignment = Alignment.Center,
-    ) {
-        Icon(
-            imageVector = Icons.Outlined.Close,
-            contentDescription = stringResource(R.string.session_task_end_process),
-            tint = GlassExitTint,
-            modifier = Modifier.size((16f * paneScale).dp),
-        )
-    }
-}
-
-@Composable
-private fun PaneSectionLabel(text: String) {
+internal fun PaneSectionLabel(text: String) {
     val paneScale = LocalPaneScale.current
     Text(
         text = text,
@@ -5360,81 +2390,6 @@ private fun PaneSectionLabel(text: String) {
         fontWeight = FontWeight.Medium,
         letterSpacing = 0.3.sp,
     )
-}
-
-@Composable
-private fun GyroscopeActivatorDropdown(
-    currentLabel: String,
-    onSelected: (Int) -> Unit,
-) {
-    val paneScale = LocalPaneScale.current
-    val names = stringArrayResource(R.array.button_options)
-    val keycodes = integerArrayResource(R.array.button_keycodes)
-    var expanded by remember { mutableStateOf(false) }
-    val parentNav = LocalPaneNav.current
-    val optionRegistry = remember { PaneNavRegistry() }
-    LaunchedEffect(expanded) {
-        if (expanded) {
-            optionRegistry.reset()
-            optionRegistry.controllerActive = true
-            parentNav?.overlay = optionRegistry
-            parentNav?.overlayClose = { expanded = false }
-        } else if (parentNav?.overlay === optionRegistry) {
-            parentNav.overlay = null
-            parentNav.overlayClose = null
-        }
-    }
-
-    val cornerRadius = (14f * paneScale).dp
-    val shape = RoundedCornerShape(cornerRadius)
-    val interactionSource = remember { MutableInteractionSource() }
-    val pressed = interactionSource.collectIsPressedAsState().value
-    val bgColor by animateColorAsState(
-        targetValue = if (pressed) PaneInnerPressed else PaneInnerResting,
-        animationSpec = tween(140),
-        label = "gyroActivatorDropdownBg",
-    )
-
-    Box(modifier = Modifier.fillMaxWidth()) {
-        Row(
-            modifier =
-                Modifier
-                    .fillMaxWidth()
-                    .clip(shape)
-                    .background(bgColor)
-                    .border(1.dp, RestingCardBorder, shape)
-                    .paneNavItem(cornerRadius = cornerRadius, onActivate = { expanded = true })
-                    .clickable(
-                        interactionSource = interactionSource,
-                        indication = null,
-                    ) { expanded = true }
-                    .padding(horizontal = (12f * paneScale).dp, vertical = (10f * paneScale).dp),
-            verticalAlignment = Alignment.CenterVertically,
-        ) {
-            Text(
-                text = currentLabel,
-                color = DrawerTextPrimary,
-                fontSize = (14f * paneScale).sp,
-                fontWeight = FontWeight.Medium,
-                modifier = Modifier.weight(1f),
-            )
-            Icon(
-                imageVector = Icons.Outlined.ArrowDropDown,
-                contentDescription = null,
-                tint = DrawerTextSecondary,
-                modifier = Modifier.size((22f * paneScale).dp),
-            )
-        }
-
-        InputControlsOptionsPopup(
-            expanded = expanded,
-            options = names.toList(),
-            selectedIndex = names.indexOfFirst { it == currentLabel }.coerceAtLeast(0),
-            onSelected = { index -> onSelected(keycodes[index]) },
-            onDismiss = { expanded = false },
-            optionRegistry = optionRegistry,
-        )
-    }
 }
 
 @Composable
@@ -5526,7 +2481,7 @@ private fun DrawerReadOnlyValueRow(
 }
 
 @Composable
-private fun DrawerSliderRow(
+internal fun DrawerSliderRow(
     label: String,
     valueText: String,
     value: Float,
@@ -5583,7 +2538,7 @@ private fun DrawerSliderRow(
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
-private fun CompactSlider(
+internal fun CompactSlider(
     value: Float,
     onValueChange: (Float) -> Unit,
     valueRange: ClosedFloatingPointRange<Float>,
@@ -5650,14 +2605,14 @@ private fun CompactSlider(
     }
 }
 
-private fun Float.snapToStep(
+internal fun Float.snapToStep(
     step: Float,
     min: Float,
     max: Float,
 ): Float = (min + (((this - min) / step).roundToInt() * step)).coerceIn(min, max)
 
 @Composable
-private fun DialogFocusButton(
+internal fun DialogFocusButton(
     label: String,
     textColor: Color,
     backgroundColor: Color,
@@ -5688,186 +2643,7 @@ private fun DialogFocusButton(
 }
 
 @Composable
-private fun HUDMetricInputDialog(
-    editor: HUDMetricEditor,
-    initialPercent: Int,
-    onDismiss: () -> Unit,
-    onConfirm: (Int) -> Unit,
-) {
-    var value by remember { mutableStateOf(initialPercent.toString()) }
-    val keyboardController = LocalSoftwareKeyboardController.current
-
-    fun submit() {
-        val parsed = value.toIntOrNull() ?: initialPercent
-        onConfirm(parsed.coerceIn(editor.minPercent, editor.maxPercent))
-    }
-
-    val focusRequester = remember { FocusRequester() }
-    WinNativeDialogShell(
-        onDismiss = onDismiss,
-        title =
-            when (editor) {
-                HUDMetricEditor.ALPHA -> stringResource(R.string.session_drawer_hud_alpha_input_title)
-                HUDMetricEditor.BACKGROUND_ALPHA -> stringResource(R.string.session_drawer_hud_background_alpha_input_title)
-                HUDMetricEditor.SCALE -> stringResource(R.string.session_drawer_hud_scale_input_title)
-            },
-        maxWidth = 380.dp,
-    ) {
-      LaunchedEffect(Unit) { runCatching { focusRequester.requestFocus() } }
-      Column(
-          modifier =
-              Modifier
-                  .controllerMenuInput(
-                      onDismiss = onDismiss,
-                      onStart = {
-                          keyboardController?.hide()
-                          submit()
-                      },
-                  ),
-      ) {
-        Text(
-            text = stringResource(R.string.session_drawer_hud_input_hint, editor.minPercent, editor.maxPercent),
-            color = DrawerTextSecondary,
-            fontSize = 13.sp,
-            lineHeight = 18.sp,
-        )
-        Spacer(Modifier.height(14.dp))
-        OutlinedTextField(
-            value = value,
-            onValueChange = { incoming -> value = incoming.filter(Char::isDigit) },
-            modifier = Modifier.fillMaxWidth(),
-            singleLine = true,
-            suffix = {
-                Text(
-                    text = "%",
-                    color = DrawerTextSecondary,
-                    fontSize = 13.sp,
-                )
-            },
-            textStyle = androidx.compose.material3.MaterialTheme.typography.bodyMedium.copy(color = DrawerTextPrimary),
-            colors =
-                OutlinedTextFieldDefaults.colors(
-                    focusedBorderColor = DrawerAccent,
-                    unfocusedBorderColor = DrawerOutline,
-                    focusedTextColor = DrawerTextPrimary,
-                    unfocusedTextColor = DrawerTextPrimary,
-                    focusedContainerColor = DrawerBackground,
-                    unfocusedContainerColor = DrawerBackground,
-                    focusedLabelColor = DrawerTextSecondary,
-                    unfocusedLabelColor = DrawerTextSecondary,
-                    cursorColor = DrawerAccent,
-                ),
-            keyboardOptions =
-                KeyboardOptions(
-                    keyboardType = KeyboardType.Number,
-                    imeAction = ImeAction.Done,
-                ),
-            keyboardActions =
-                KeyboardActions(
-                    onDone = {
-                        keyboardController?.hide()
-                        submit()
-                    },
-                ),
-        )
-        Spacer(Modifier.height(16.dp))
-        Box(
-            modifier =
-                Modifier
-                    .fillMaxWidth()
-                    .height(1.dp)
-                    .background(DrawerOutline),
-        )
-        Spacer(Modifier.height(16.dp))
-        Row(
-            modifier = Modifier.fillMaxWidth(),
-            horizontalArrangement = Arrangement.spacedBy(10.dp, Alignment.End),
-        ) {
-            DialogFocusButton(
-                label = stringResource(R.string.common_ui_cancel),
-                textColor = DrawerTextPrimary,
-                backgroundColor = PaneInnerResting,
-                borderColor = RestingCardBorder,
-                onClick = onDismiss,
-            )
-            DialogFocusButton(
-                label = stringResource(R.string.common_ui_apply),
-                textColor = DrawerAccent,
-                backgroundColor = DrawerAccent.copy(alpha = 0.12f),
-                borderColor = DrawerAccent.copy(alpha = 0.3f),
-                focusRequester = focusRequester,
-                onClick = {
-                    keyboardController?.hide()
-                    submit()
-                },
-            )
-        }
-      }
-    }
-}
-
-@Composable
-private fun HUDToggleChip(
-    label: String,
-    checked: Boolean,
-    onClick: () -> Unit,
-    modifier: Modifier = Modifier,
-) {
-    val paneScale = LocalPaneScale.current
-    val interactionSource = remember { MutableInteractionSource() }
-    val pressed = interactionSource.collectIsPressedAsState().value
-    val bgColor by animateColorAsState(
-        targetValue =
-            when {
-                pressed -> PaneInnerPressed
-                else -> PaneInnerResting
-            },
-        animationSpec = tween(140),
-        label = "hudChipBg",
-    )
-    val borderColor by animateColorAsState(
-        targetValue = if (checked) DrawerAccent else RestingCardBorder,
-        animationSpec = tween(140),
-        label = "hudChipBorder",
-    )
-    val cornerRadius = (12f * paneScale).dp
-    val shape = RoundedCornerShape(cornerRadius)
-    val indicatorSize = (10f * paneScale).dp
-
-    Row(
-        modifier =
-            modifier
-                .clip(shape)
-                .background(bgColor)
-                .border(1.dp, borderColor, shape)
-                .clickable(
-                    interactionSource = interactionSource,
-                    indication = null,
-                    onClick = onClick,
-                ).padding(horizontal = (10f * paneScale).dp, vertical = (9f * paneScale).dp),
-        verticalAlignment = Alignment.CenterVertically,
-    ) {
-        Box(
-            modifier =
-                Modifier
-                    .size(indicatorSize)
-                    .clip(CircleShape)
-                    .background(if (checked) DrawerAccent else Color(0x14FFFFFF)),
-        )
-        Spacer(Modifier.width((8f * paneScale).dp))
-        Text(
-            text = label,
-            color = DrawerTextPrimary,
-            fontSize = (13f * paneScale).sp,
-            fontWeight = if (checked) FontWeight.SemiBold else FontWeight.Medium,
-            maxLines = 1,
-            overflow = TextOverflow.Ellipsis,
-        )
-    }
-}
-
-@Composable
-private fun DrawerBooleanRow(
+internal fun DrawerBooleanRow(
     title: String,
     checked: Boolean,
     onCheckedChange: (Boolean) -> Unit,
@@ -5941,282 +2717,14 @@ private fun DrawerBooleanRow(
     }
 }
 
-private val RECORD_QUALITY_LABELS = listOf("Performance", "Balance", "Quality")
+internal val RECORD_QUALITY_LABELS = listOf("Performance", "Balance", "Quality")
 
-/** Centered popup for choosing recording fps / resolution / quality (+ Record UI), then Record Now. */
-@Composable
-private fun RecordSettingsDialog(
-    config: RecordUiConfig,
-    onDismiss: () -> Unit,
-    onRecordNow: (fpsIndex: Int, resolutionIndex: Int, quality: Int, recordUI: Boolean) -> Unit,
-) {
-    val fpsOptions = config.fpsOptions.ifEmpty { listOf(60) }
-    val resOptions = config.resolutionLabels.ifEmpty { listOf("Native") }
-
-    var fpsIndex by remember { mutableStateOf(config.fpsIndex.coerceIn(0, fpsOptions.lastIndex)) }
-    var resIndex by remember { mutableStateOf(config.resolutionIndex.coerceIn(0, resOptions.lastIndex)) }
-    var quality by remember { mutableStateOf(config.quality.coerceIn(0, RECORD_QUALITY_LABELS.lastIndex)) }
-    var recordUI by remember { mutableStateOf(config.recordUI) }
-
-    val recordNav = remember { SharedPaneNavRegistry() }
-    val doRecord = { onRecordNow(fpsIndex, resIndex, quality, recordUI) }
-
-    val shape = RoundedCornerShape(16.dp)
-    // Cap card height (landscape is short); settings scroll, the Record Now button stays pinned.
-    val maxCardHeight = (LocalConfiguration.current.screenHeightDp * 0.92f).dp
-    Dialog(
-        onDismissRequest = onDismiss,
-        properties = DialogProperties(usePlatformDefaultWidth = false, decorFitsSystemWindows = false),
-    ) {
-        CompositionLocalProvider(SharedLocalPaneNav provides recordNav) {
-        DialogPaneNav(recordNav, onDismiss = onDismiss, onStart = doRecord)
-        Box(
-            modifier = Modifier.fillMaxSize().safeDrawingPadding().padding(horizontal = 14.dp, vertical = 8.dp),
-            contentAlignment = Alignment.Center,
-        ) {
-            Column(
-                modifier =
-                    Modifier
-                        .widthIn(max = 360.dp)
-                        .fillMaxWidth()
-                        .heightIn(max = maxCardHeight)
-                        .clip(shape)
-                        .background(PaneSurfaceColor)
-                        .border(1.dp, RestingCardBorder, shape)
-                        .padding(horizontal = 16.dp, vertical = 14.dp),
-                verticalArrangement = Arrangement.spacedBy(12.dp),
-            ) {
-                Row(
-                    modifier = Modifier.fillMaxWidth(),
-                    verticalAlignment = Alignment.CenterVertically,
-                    horizontalArrangement = Arrangement.spacedBy(8.dp),
-                ) {
-                    Icon(
-                        imageVector = Icons.Outlined.FiberManualRecord,
-                        contentDescription = null,
-                        tint = RecordRed,
-                        modifier = Modifier.size(18.dp),
-                    )
-                    Text(
-                        text = stringResource(R.string.session_record_settings_title),
-                        color = DrawerTextPrimary,
-                        fontSize = 15.sp,
-                        fontWeight = FontWeight.SemiBold,
-                    )
-                }
-
-                Column(
-                    modifier = Modifier.weight(1f, fill = false).verticalScroll(rememberScrollState()),
-                    verticalArrangement = Arrangement.spacedBy(12.dp),
-                ) {
-                    Box(
-                        modifier = Modifier.fillMaxWidth().sharedPaneNavItem(
-                            onAdjust = { dir -> if (fpsOptions.size > 1) fpsIndex = (fpsIndex + dir).coerceIn(0, fpsOptions.lastIndex) },
-                        ),
-                    ) {
-                        DrawerSliderRow(
-                            label = stringResource(R.string.session_record_fps),
-                            valueText = "${fpsOptions[fpsIndex]} fps",
-                            value = fpsIndex.toFloat(),
-                            valueRange = 0f..(fpsOptions.lastIndex.coerceAtLeast(1)).toFloat(),
-                            steps = (fpsOptions.size - 2).coerceAtLeast(0),
-                            onValueChange = { if (fpsOptions.size > 1) fpsIndex = it.roundToInt().coerceIn(0, fpsOptions.lastIndex) },
-                        )
-                    }
-
-                    Box(
-                        modifier = Modifier.fillMaxWidth().sharedPaneNavItem(
-                            onAdjust = { dir -> if (resOptions.size > 1) resIndex = (resIndex + dir).coerceIn(0, resOptions.lastIndex) },
-                        ),
-                    ) {
-                        DrawerSliderRow(
-                            label = stringResource(R.string.session_record_resolution),
-                            valueText = resOptions[resIndex],
-                            value = resIndex.toFloat(),
-                            valueRange = 0f..(resOptions.lastIndex.coerceAtLeast(1)).toFloat(),
-                            steps = (resOptions.size - 2).coerceAtLeast(0),
-                            onValueChange = { if (resOptions.size > 1) resIndex = it.roundToInt().coerceIn(0, resOptions.lastIndex) },
-                        )
-                    }
-
-                    Box(
-                        modifier = Modifier.fillMaxWidth().sharedPaneNavItem(
-                            onAdjust = { dir -> quality = (quality + dir).coerceIn(0, RECORD_QUALITY_LABELS.lastIndex) },
-                        ),
-                    ) {
-                        DrawerSliderRow(
-                            label = stringResource(R.string.session_record_quality),
-                            valueText = RECORD_QUALITY_LABELS[quality],
-                            value = quality.toFloat(),
-                            valueRange = 0f..(RECORD_QUALITY_LABELS.lastIndex).toFloat(),
-                            steps = (RECORD_QUALITY_LABELS.size - 2).coerceAtLeast(0),
-                            onValueChange = { quality = it.roundToInt().coerceIn(0, RECORD_QUALITY_LABELS.lastIndex) },
-                        )
-                    }
-
-                    Box(
-                        modifier = Modifier.fillMaxWidth().sharedPaneNavItem(
-                            onActivate = { recordUI = !recordUI },
-                        ),
-                    ) {
-                        DrawerBooleanRow(
-                            title = stringResource(R.string.session_record_include_ui),
-                            checked = recordUI,
-                            onCheckedChange = { recordUI = it },
-                            subtitle = stringResource(R.string.session_record_include_ui_subtitle),
-                        )
-                    }
-                }
-
-                Button(
-                    onClick = doRecord,
-                    modifier = Modifier.fillMaxWidth().height(48.dp).sharedPaneNavItem(
-                        cornerRadius = 12.dp,
-                        onActivate = doRecord,
-                        isEntry = true,
-                    ),
-                    shape = RoundedCornerShape(12.dp),
-                    colors =
-                        ButtonDefaults.buttonColors(
-                            containerColor = RecordRed,
-                            contentColor = Color.White,
-                        ),
-                ) {
-                    Icon(
-                        imageVector = Icons.Outlined.FiberManualRecord,
-                        contentDescription = null,
-                        tint = Color.White,
-                        modifier = Modifier.size(18.dp),
-                    )
-                    Spacer(Modifier.width(8.dp))
-                    Text(
-                        text = stringResource(R.string.session_record_now),
-                        color = Color.White,
-                        fontSize = 15.sp,
-                        fontWeight = FontWeight.Bold,
-                    )
-                }
-            }
-        }
-        }
-    }
-}
-
-private const val FPS_LIMITER_MIN = 30
-private const val FPS_LIMITER_DEFAULT = 60
-
-@Composable
-private fun FPSLimiterCard(
-    currentLimit: Int,
-    maxRefreshRate: Int,
-    onLimitChanged: (Int) -> Unit,
-) {
-    val paneScale = LocalPaneScale.current
-    val enabled = currentLimit > 0
-    val maxFps = maxRefreshRate.coerceAtLeast(FPS_LIMITER_MIN)
-    val steps = (maxFps - FPS_LIMITER_MIN - 1).coerceAtLeast(0)
-
-    // Slider position tracked locally (readout follows the drag, value survives an off/on toggle); the commit is deferred to release and re-seeds when maxFps changes (e.g. a mid-game refresh-rate change that clamps the limit).
-    var sliderValue by remember(maxFps) {
-        mutableStateOf(
-            (if (currentLimit > 0) currentLimit else FPS_LIMITER_DEFAULT)
-                .coerceIn(FPS_LIMITER_MIN, maxFps)
-                .toFloat(),
-        )
-    }
-
-    LaunchedEffect(currentLimit) {
-        if (currentLimit > 0) {
-            val target = currentLimit.coerceIn(FPS_LIMITER_MIN, maxFps).toFloat()
-            if (target != sliderValue) sliderValue = target
-        }
-    }
-
-    val borderColor by animateColorAsState(
-        targetValue = if (enabled) ActiveCardBorder else RestingCardBorder,
-        animationSpec = tween(140),
-        label = "fpsLimiterCardBorder",
-    )
-    val shape = RoundedCornerShape((14f * paneScale).dp)
-
-    Column(
-        modifier =
-            Modifier
-                .fillMaxWidth()
-                .clip(shape)
-                .background(PaneInnerResting)
-                .border(1.dp, borderColor, shape)
-                .padding(horizontal = (12f * paneScale).dp, vertical = (8f * paneScale).dp),
-    ) {
-        Row(
-            modifier =
-                Modifier
-                    .fillMaxWidth()
-                    .clickable(
-                        interactionSource = remember { MutableInteractionSource() },
-                        indication = null,
-                    ) { onLimitChanged(if (enabled) 0 else sliderValue.roundToInt()) },
-            verticalAlignment = Alignment.CenterVertically,
-        ) {
-            Column(modifier = Modifier.weight(1f)) {
-                Text(
-                    text = stringResource(R.string.session_drawer_fps_limiter),
-                    color = DrawerTextPrimary,
-                    fontSize = (14f * paneScale).sp,
-                    fontWeight = FontWeight.Medium,
-                )
-                Spacer(Modifier.height(2.dp))
-                Text(
-                    text =
-                        if (enabled) {
-                            "${sliderValue.roundToInt()} FPS"
-                        } else {
-                            stringResource(R.string.session_drawer_fps_limiter_off)
-                        },
-                    color = if (enabled) DrawerAccent else DrawerTextSecondary,
-                    fontSize = (12f * paneScale).sp,
-                    fontWeight = if (enabled) FontWeight.SemiBold else FontWeight.Normal,
-                )
-            }
-            CompositionLocalProvider(LocalRippleConfiguration provides null) {
-                Switch(
-                    checked = enabled,
-                    onCheckedChange = { on -> onLimitChanged(if (on) sliderValue.roundToInt() else 0) },
-                    colors = outlinedSwitchColors(DrawerAccent, DrawerTextSecondary),
-                )
-            }
-        }
-
-        AnimatedVisibility(
-            visible = enabled,
-            enter =
-                expandVertically(
-                    animationSpec = tween(durationMillis = 220, easing = FastOutSlowInEasing),
-                    expandFrom = Alignment.Top,
-                ) + fadeIn(animationSpec = tween(durationMillis = 160, easing = FastOutSlowInEasing)),
-            exit =
-                shrinkVertically(
-                    animationSpec = tween(durationMillis = 180, easing = FastOutSlowInEasing),
-                    shrinkTowards = Alignment.Top,
-                ) + fadeOut(animationSpec = tween(durationMillis = 120, easing = FastOutSlowInEasing)),
-        ) {
-            Column {
-                Spacer(Modifier.height((6f * paneScale).dp))
-                CompactSlider(
-                    value = sliderValue,
-                    onValueChange = { sliderValue = it },
-                    valueRange = FPS_LIMITER_MIN.toFloat()..maxFps.toFloat(),
-                    steps = steps,
-                    onValueChangeFinished = { onLimitChanged(sliderValue.roundToInt()) },
-                )
-            }
-        }
-    }
-}
+internal const val FPS_LIMITER_MIN = 30
+internal const val FPS_LIMITER_DEFAULT = 60
 
 @OptIn(ExperimentalLayoutApi::class)
 @Composable
-private fun ChipFlow(content: @Composable () -> Unit) {
+internal fun ChipFlow(content: @Composable () -> Unit) {
     val paneScale = LocalPaneScale.current
     val gap = (8f * paneScale).dp
     FlowRow(
